% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_intercept}
\alias{add_intercept}
\title{Adds an intercept to a matrix}
\usage{
add_intercept(mat)
}
\arguments{
\item{mat}{Any matrix.}
}
\value{
Matrix with intercept.
}
\description{
\code{\link{add_intercept}} adds an intercept to a matrix.
}
\examples{
X = matrix(c(0.2, 0.3, 0.4, 0.5, 0.6, 0.7), ncol = 2, dimnames = list(1:3, c("a", "b")))
X
X_with_intercept = add_intercept(X)
X_with_intercept

}
