% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agg_array}
\alias{agg_array}
\title{Aggregate 3D array}
\usage{
agg_array(a, w)
}
\arguments{
\item{a}{A 3D array}

\item{w}{A numeric vector.
The length of w should match the third dimension of a.}
}
\value{
A matrix of the same dimension as the first two dimensions of a.
}
\description{
This function computes the weighted sum along the third dimension of a
three dimensional array.
}
\keyword{internal}
