% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_final_method}
\alias{create_final_method}
\title{Method creation for ensemble (extended version)}
\usage{
create_final_method(method, NuPa, K)
}
\arguments{
\item{method}{Either:
1. A list of methods (like in create_method) that will be used for all nuisance parameters, or
2. A named list where each name corresponds to a nuisance parameter and contains a list of methods for that parameter}

\item{NuPa}{Character vector of nuisance parameters to be estimated}

\item{K}{Number of unique treatment statuses (default = 2, binary treatment)}
}
\value{
Nested list object organized by nuisance parameter that can be passed as input to \code{\link{ensemble}}
}
\description{
Creates the methods to be used in the subsequent \code{\link{ensemble}} model,
allowing for either a simple method list or nuisance-parameter-specific method lists.
}
\keyword{internal}
