% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Second observational dataset from the Maren's thesis}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4424 rows and 37 columns.
}
\usage{
dataset
}
\description{
The dataset contains data on enrollment information, demographics, 
socio-economic factors, and academic performance after the first and second 
semesters. This dataset is used to develop classification models predicting 
student dropout and academic success. The classification task is formulated 
with three categories, with a significant imbalance toward one class. The 
dataset has 4424 observations and 36 confounding variables 
(102 after encoding the categorical columns).
}
\keyword{datasets}
