% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble learner training}
\usage{
ensemble(ml, x, y, nfolds = 5, quiet = TRUE)
}
\arguments{
\item{ml}{List of methods built via \code{\link{create_method}} to be used in
ensemble model}

\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}

\item{nfolds}{Number of folds used in cross-validation of ensemble weights
(default \code{nfolds = 5})}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}
}
\value{
Ensemble object containing:
\item{ml_fit}{ensemble_core object fitted on whole training sample}
\item{nnls_weights}{the weights that each method receives in the ensemble}
\item{mse_cv}{cross-validated MSEs for each method}
\item{fit_cv}{matrix containing the cross-validated predictions for each method}
\item{ml}{list of raw methods built via \code{\link{create_method}}}

 @export
}
\description{
\code{\link{ensemble}} trains an ensemble learner on the given training data.
}
