% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_core}
\alias{ensemble_core}
\title{Ensemble core fitting}
\usage{
ensemble_core(
  methods,
  X_tr,
  Y_tr,
  quiet = TRUE,
  pb = NULL,
  pb_cf = NULL,
  pb_cv = NULL,
  pb_np = NULL
)
}
\arguments{
\item{methods}{List of methods created with \code{\link{create_method}} to be 
used in the ensemble model.}

\item{X_tr}{Covariate matrix of the training sample.}

\item{Y_tr}{Vector of outcomes of the training sample.}

\item{quiet}{Logical. If \code{FALSE}, progress output is printed to the console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_cf}{Integer indicating the current cross-fitting fold 
(for progress bar updates).}

\item{pb_cv}{Integer indicating the current cross-validation fold 
(for progress bar updates).}

\item{pb_np}{String indicating the current nuisance parameter being estimated 
or predicted (for progress bar updates).}
}
\value{
An object of class \code{"ensemble_core"} containing the 
 trained models of the ensemble.
}
\description{
Train the inputted machine learning methods for one given set of covariates 
and outcomes within the ensemble framework.
}
\keyword{internal}
