% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_short.R
\name{ensemble_short}
\alias{ensemble_short}
\title{Ensemble learner training for short stacking}
\usage{
ensemble_short(
  method,
  X,
  Y,
  cf_mat,
  subset = NULL,
  storeModels = c("No", "Memory"),
  quiet = TRUE,
  pb = NULL,
  pb_np = NULL
)
}
\arguments{
\item{method}{List of methods built via \code{\link{create_method}} to be used in
ensemble model.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds.}

\item{subset}{Logical vector indicating which observations to use for determining
ensemble weights. If not provided, all observations are used.}

\item{storeModels}{Specifies whether to store the models. Value has to be an item
from \code{c("No", "Memory")} with "No" being the default value.}

\item{quiet}{If FALSE, progress output is printed into console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_np}{String indicating the current nuisance parameter being
estimated/predicted (for progress bar updates).}
}
\value{
List object containing:
\item{fit_cv}{matrix of dimension \code{nrow(X)} X \code{length(method)} containing
the cross-fitted predictions of the machine learning methods from the ensemble}
\item{method}{list of fitted machine learning models from all cross-fitting folds
(if \code{storeModels} is not set to "No")}
}
\description{
\code{\link{ensemble_short}} trains an ensemble learner on the given training data
in the spirit of a short-stacking procedure.
}
\keyword{internal}
