% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{forest_drf_fit}
\alias{forest_drf_fit}
\title{Fit distributional random forest model}
\usage{
forest_drf_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{arguments}{List of arguments passed to \code{\link[drf]{drf}}.}
}
\value{
An object of class \code{forest_drf_fit} containing the fitted
  \code{\link[drf]{drf}} model.
}
\description{
Fits a distributional random forest using the \code{drf} package with
Fourier MMD splitting rule.
}
\keyword{internal}
