% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{forest_grf_fit}
\alias{forest_grf_fit}
\title{Fits Random Forest}
\usage{
forest_grf_fit(x, y, args = list())
}
\arguments{
\item{x}{Matrix of covariates}

\item{y}{vector of outcomes}

\item{args}{List of arguments passed to \code{\link[grf]{regression_forest}}}
}
\value{
An object with S3 class \code{\link[grf]{regression_forest}}
}
\description{
\code{\link{forest_grf_fit}} fits a random forest using the
\code{\link{grf}} package.
}
\keyword{internal}
