% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_learners.r
\name{forest_grf_tune}
\alias{forest_grf_tune}
\title{Tune \code{grf} Random Forest Hyperparameters}
\usage{
forest_grf_tune(X, Y)
}
\arguments{
\item{X}{Covariate matrix or data frame.}

\item{Y}{Numeric vector containing the outcome variable.}
}
\value{
List containing tuned hyperparameters
}
\description{
Performs default \code{grf} hyperparameter tuning.
}
\keyword{internal}
