% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_weights}
\alias{format_weights}
\title{Format ensemble weights for short- or standard-stacking}
\usage{
format_weights(ens_weights, cv)
}
\arguments{
\item{ens_weights}{List of ensemble weights to be formatted}

\item{cv}{Integer indicating number of cross-validation folds}
}
\value{
A data.frame (\code{cv = 1}) or list of data.frames (\code{cv > 1}) with formatted weights
}
\description{
Format ensemble weights for short- or standard-stacking
}
\keyword{internal}
