% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{get_outcome_weights}
\alias{get_outcome_weights}
\title{Extraction of smoother weights for ensemble learner}
\usage{
get_outcome_weights(np_object, NuPa, quiet = TRUE, subset = NULL, ...)
}
\arguments{
\item{np_object}{Stacked ensemble learner from \code{\link{ensemble_short}} or \code{\link{ensemble}}.}

\item{NuPa}{String vector specifying which nuisance parameters to extract smoother matrices for.}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}

\item{subset}{Logical vector indicating which observations to use for determining
ensemble weights. If not provided, all observations are used.}

\item{...}{Ignore unused arguments}

\item{ml}{List of ML models by \code{\link{create_method}} that was used as
iobjectut for \code{\link{ensemble_short}} or \code{\link{ensemble}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{d_mat}{Logical matrix of treatment indicators (N X T+1). For example created by \code{\link{prep_w_mat}}.}

\item{z_mat}{Logical matrix of instrument indicators (N X T+1). For example created by \code{\link{prep_w_mat}}.}

\item{cv}{Number of cross-validation folds when estimating ensemble model.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}
}
\value{
Matrix of dimension \code{nrow(X)} X \code{nrow(X)} containing
ensemble smoother weights, or list of such matrices if multiple models are requested.
}
\description{
Extract smoother weights from short- or standard-stacked ensemble model 
created by \code{\link{ensemble_short}} or \code{\link{ensemble}}.
}
