% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother.R
\name{get_smoother}
\alias{get_smoother}
\title{Extraction of smoother weights for ensemble learner}
\usage{
get_smoother(
  object,
  ml,
  x,
  y,
  subset = NULL,
  nnls_w = NULL,
  cv = cv,
  cf_mat,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{object}{Stacked ensemble learner from \code{\link{ensemble_short}} or \code{\link{ensemble}}.}

\item{ml}{List of ML models by \code{\link{create_method}} that was used as
input for \code{\link{ensemble_short}} or \code{\link{ensemble}}.}

\item{x}{Covariate matrix of training sample.}

\item{y}{Vector of outcomes of training sample.}

\item{subset}{Logical vector indicating which observations to use for determining
ensemble weights. If not provided, all observations are used.}

\item{nnls_w}{Ensemble weights to aggregate predictions from different learners (optional).
Needs to be a vector of length \code{ncol(object$fit_cv)}.}

\item{cv}{Number of cross-validation folds when estimating ensemble model.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of dimension \code{nrow(x)} x \code{nrow(x)} containing
ensemble smoother weights.
}
\description{
Extract smoother weights from short- or standard-stacked ensemble model 
created by \code{\link{ensemble_short}} or \code{\link{ensemble}}.
}
