% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother_weights}
\alias{get_smoother_weights}
\title{Smoother weights for outcome models}
\usage{
get_smoother_weights(object, NuPa, quiet = TRUE, subset = NULL, ...)
}
\arguments{
\item{object}{A \code{NuisanceParameters} object created by
\code{\link{nuisance_parameters}}, or a path to an RDS file containing one.}

\item{NuPa}{Character vector specifying which nuisance parameters to extract 
smoothers for. Available outcome nuisance parameters 
are \code{c("Y.hat", "Y.hat.d", "Y.hat.z")}.}

\item{quiet}{If \code{FALSE}, the method currently computed is printed to the console.}

\item{subset}{Logical vector indicating which observations to use when determining
ensemble weights. If not provided, all observations are used.}

\item{...}{Additional arguments.}
}
\value{
A list of matrices of dimension \code{nrow(X)} × \code{nrow(X)} containing ensemble smoother
  matrices
}
\description{
Extract smoother matrices for nuisance parameter outcome models from a
\code{NuisanceParameters} object created by \code{\link{nuisance_parameters}}.
}
