% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{glm_fit}
\alias{glm_fit}
\title{Fit binary logistic or probit regression using \code{glm}}
\usage{
glm_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of binary outcomes.}

\item{arguments}{List of arguments passed to \code{stats::glm}. To use probit,
specify \code{family = binomial(link = "probit")}. Default is logit.}
}
\value{
An object of class \code{glm} containing the fitted model.
}
\description{
Fits binary regression models using \code{glm}. Supports both logistic
 (logit) and probit link functions. For binary outcomes only.
}
\keyword{internal}
