% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{knn_fit}
\alias{knn_fit}
\title{Pseudo-fitting for k-Nearest Neighbors}
\usage{
knn_fit(arguments = list(), ...)
}
\arguments{
\item{arguments}{List of arguments to be used in k-NN prediction. Typically
includes \code{k} (number of neighbors).}

\item{...}{Ignored additional arguments.}
}
\value{
A list of class \code{knn_fit} containing the arguments for k-NN.
}
\description{
Creates a k-NN model object without actual fitting, as k-NN fitting occurs
naturally during prediction.
}
\keyword{internal}
