% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{knn_prop_fit}
\alias{knn_prop_fit}
\title{Calculates k-Nearest Neighbor model using the \code{\link{kknn}} package}
\usage{
knn_prop_fit(X, Y, arguments = list(kmax = floor(0.05 * nrow(X))))
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{vector of outcomes}

\item{arguments}{List of arguments passed to  \code{\link{kknn}}}
}
\value{
An object with S3 class "kknn"
}
\description{
Calculates k-Nearest Neighbor model using the \code{\link{kknn}} package
}
\keyword{internal}
