% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{lasso_fit}
\alias{lasso_fit}
\title{Fits Lasso regression}
\usage{
lasso_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{Y}{vector of outcomes}

\item{arguments}{List of arguments passed to \code{\link[glmnet]{cv.glmnet}}}
}
\value{
An object with S3 class \code{\link[glmnet]{cv.glmnet}}
}
\description{
\code{\link{lasso_fit}} estimates cross-validated Lasso regression based on
the \code{\link{glmnet}} package.
}
\keyword{internal}
