% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{logit_fit}
\alias{logit_fit}
\title{This function calculates the (multinomial) logistic regression based on the \code{\link{glmnet}} package}
\usage{
logit_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{arguments}{List of arguments passed to  \code{\link{glmnet}}}
}
\value{
An object with S3 class "glmnet"
}
\description{
This function calculates the (multinomial) logistic regression based on the \code{\link{glmnet}} package
}
