% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{logit_nnet_fit}
\alias{logit_nnet_fit}
\title{This function calculates the (multinomial) logistic regression based on the \code{\link{nnet}} package}
\usage{
logit_nnet_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{arguments}{List of arguments passed to  \code{\link{nnet}}}
}
\value{
An object with S3 method for class 'nnet'
}
\description{
This function calculates the (multinomial) logistic regression based on the \code{\link{nnet}} package
}
