% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cf_preds}
\alias{make_cf_preds}
\title{Prepare \code{cf_preds} matrix for cross-fitted ensemble predictions}
\usage{
make_cf_preds(methods, N, Y)
}
\arguments{
\item{methods}{List of methods built via \code{\link{create_method}} to be used in ensemble}

\item{N}{Number of observations}

\item{Y}{Outcome vector (used to determine number of classes for multinomial outcomes)}
}
\value{
A matrix of NAs with dimensions \code{N} x \code{length(methods)} for binary outcomes,
        or a 3D array with dimensions \code{N} x \code{K} x \code{length(methods)} for multinomial
        outcomes (where K is number of unique classes in Y)
}
\description{
Creates a matrix or array to store cross-fitted ensemble predictions.
}
\keyword{internal}
