% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{mean_fit}
\alias{mean_fit}
\title{Arithmetic mean}
\usage{
mean_fit(X, Y, ...)
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{...}{Ignore unused arguments}
}
\value{
Returns list containing mean and number of observations.
}
\description{
\code{\link{mean_fit}} acts as prediction function by calculating the
arithmetic mean of the training outcomes.
}
\keyword{internal}
