% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLeffect.R
\name{MLeffect}
\alias{MLeffect}
\title{Double ML estimators (binary treatments only)}
\usage{
MLeffect(
  Y,
  D,
  X,
  Z = NULL,
  NuPa.hat,
  estimators = c("PLR", "PLR_IV", "AIPW_ATE", "Wald_AIPW")
)
}
\arguments{
\item{Y}{Numeric vector containing the outcome variable.}

\item{D}{Optional binary treatment variable.}

\item{X}{Covariate matrix with N rows and p columns.}

\item{Z}{Optional binary instrumental variable.}

\item{NuPa.hat}{a list of estimated nuisance parameters from \code{\link{nuisance_parameters}}}

\item{estimators}{String (vector) indicating which estimators should be run.
Current menu: c("PLR","PLR_IV","AIPW_ATE","Wald_AIPW")}
}
\value{
An object of class `dml_with_smoother` containing:
\itemize{
  \item \code{results} - A list of estimated target parameters (PLR, PLR_IV, AIPW_ATE, Wald_AIPW)
  \item \code{NuPa.hat} - A list of input nuisance parameters
  \item \code{data} - Input data (Y, D, Z, X)
  \item \code{numbers} - Sample information (N, n_estimators)
}
}
\description{
For test use only, copied from \code{OutcomeWeights} package
}
