% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{nb_bernoulli_fit}
\alias{nb_bernoulli_fit}
\title{This function calculates the bernoulli Naive Bayes model based on the \code{\link{naivebayes}} package}
\usage{
nb_bernoulli_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{arguments}{List of arguments passed to  \code{\link{naivebayes}}}
}
\value{
An object with S3 method for class 'bernoulli_naive_bayes'
}
\description{
This function calculates the bernoulli Naive Bayes model based on the \code{\link{naivebayes}} package
}
