% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{nb_gaussian_fit}
\alias{nb_gaussian_fit}
\title{Fit Gaussian Naive Bayes using \code{naivebayes}}
\usage{
nb_gaussian_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{arguments}{List of arguments passed to \code{naivebayes::gaussian_naive_bayes()}.}
}
\value{
An object of class \code{gaussian_naive_bayes} containing the fitted model.
}
\description{
Fits a Gaussian Naive Bayes model using the \code{naivebayes} package.
}
\keyword{internal}
