% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnls_weights}
\alias{nnls_weights}
\title{Non-linear least weights function}
\usage{
nnls_weights(X, y)
}
\arguments{
\item{X}{A matrix where each column represents a different predictor variable
and each row represents an observation}

\item{y}{A numeric vector of actual target values}
}
\value{
A numeric vector of non-linear least weights.
}
\description{
\code{\link{nnls_weights}} calculates the non-linear least weights on the
basis of the non-negative least squares algorithm.
}
\examples{
X = matrix(c(0.2, 0.3, 0.4, 0.5, 0.6, 0.7), ncol = 2)
y = c(0.25, 0.45, 0.65)
\donttest{nnls_w = nnls_weights(X, y)}

}
