% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_data.R
\docType{data}
\name{nuisance_data}
\alias{nuisance_data}
\title{Simulated example data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
data(nuisance_data)
}
\description{
Simulated data from a DGP with covariates matrix X, treatment vector w
and target vector y.
}
\examples{
# load data
data(nuisance_data)
# extract target, treatment and covariates from data
# y = nuisance_data[, "y", drop = FALSE]
# w = nuisance_data[, "w", drop = FALSE]
# X = nuisance_data[, 3:ncol(nuisance_data)]

}
\keyword{datasets}
