% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuisanceParameters-package.R
\docType{package}
\name{NuisanceParameters-package}
\alias{NuisanceParameters}
\alias{NuisanceParameters-package}
\title{NuisanceParameters: Nuisance Parameters Estimation}
\description{
Estimation of nuisance parameters using different prediction algorithms.
}
\author{
\strong{Maintainer}: Glaisner Stefan \email{stefan.glaisner@student.uni-tuebingen.de}

Other contributors:
\itemize{
  \item Knaus Michael C. \email{michael.knaus@uni-tuebingen.de} (\href{https://orcid.org/0000-0002-7328-1363}{ORCID}) [contributor]
}

}
\keyword{internal}
