% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuPa.R
\name{NuPa_honest_forest}
\alias{NuPa_honest_forest}
\title{Nuisance parameter estimation}
\usage{
NuPa_honest_forest(
  NuPa = c("Y.hat", "Yw.hat", "Yz.hat", "W.hat", "Wz.hat", "Z.hat"),
  ml,
  x,
  y = NULL,
  w = NULL,
  z = NULL,
  cf_mat,
  w_mat = NULL,
  z_mat = NULL,
  cv = 5,
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{NuPa}{String vector specifying the nuisance parameters to be estimated.
Currently supported: \code{c("Y.hat","Yw.hat","Yz.hat","W.hat","Wz.hat","Z.hat")}}

\item{ml}{List of methods to be used in \code{\link{ensemble}} estimation of
propensity score.
Methods can be created by \code{\link{create_method}}.}

\item{x}{Covariate matrix.}

\item{y}{Numerical vector containing the outcome variable.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}})}

\item{w_mat}{Logical matrix of treatment indicators (n x T+1). For example
created by \code{\link{prep_w_mat}}.}

\item{cv}{Number of cross-validation when estimating ensemble (default 5).}

\item{path}{Optional path to save the \code{\link{ensemble}} objects for later processing.
Saved as Ensemble_Yi where i is the number of the treatment in multiple treatment settings.}

\item{quiet}{If FALSE, method that is currently running is printed into console.}

\item{learner}{Vector of characters indicating whether to use S or T learner
or a combination of the two.}

\item{weights}{If TRUE, prediction weights of the outcome nuisance extracted
and saved (requires to provide a path).}
}
\value{
List of two lists. 
- \code{predictions} contains the requested nuisance parameters
- \code{smoothers} contains the smoother matrices of requested outcome nuisance parameters
- \code{cf_mat} Array of dimension n_reps x N x n_cf_folds storing indicators representing the folds used in estimation.
}
\description{
This function estimates different nuisance parameters (draft, copy from OutcomeWeights)
}
