% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{ols_fit}
\alias{ols_fit}
\title{Fits OLS}
\usage{
ols_fit(X, Y, ...)
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{...}{Ignore unused arguments}
}
\value{
Returns OLS coefficients
}
\description{
\code{\link{ols_fit}} fits ordinary least squares (OLS) to the given data.
}
\keyword{internal}
