% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{one_hot}
\alias{one_hot}
\title{One-Hot Encoding (numeric output)}
\usage{
one_hot(Y)
}
\arguments{
\item{Y}{A factor or character vector.}
}
\value{
A numeric matrix with one column per unique category.
}
\description{
Convert categorical vector to one-hot encoded matrix.
}
\keyword{internal}
