% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{ovo_fit}
\alias{ovo_fit}
\title{Fits One-Vs-One (OvO) binary classifiers for multi-class classification}
\usage{
ovo_fit(X, Y, method = "logit")
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{method}{Method to use for fitting binary classifiers (e.g., "logit" or "xgboost")}
}
\value{
List of fitted OvO binary classifiers
}
\description{
Fits One-Vs-One (OvO) binary classifiers for multi-class classification
}
