% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{plasso_fit}
\alias{plasso_fit}
\title{Fit post-lasso regression model}
\usage{
plasso_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{arguments}{List of arguments passed to \code{cv.plasso}.}
}
\value{
An object of class \code{plasso_fit} containing the fitted
  \code{cv.plasso} model.
}
\description{
Estimates cross-validated post-lasso regression using the \code{plasso} package.
Performs lasso selection followed by OLS on selected variables.
}
\keyword{internal}
