% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.ens.learner}
\alias{plot.ens.learner}
\title{Visualize ensemble weights}
\usage{
\method{plot}{ens.learner}(
  x,
  methods = NULL,
  legend_pos = c("center", "bottomright", "bottom", "bottomleft", "left", "topleft",
    "top", "topright", "right"),
  legend_size = 1,
  ...
)
}
\arguments{
\item{x}{Object containing predictions by ensemble learner (fit_cv)
and the cross-validated ensemble weights.}

\item{methods}{Vector of method names. In order of initial \code{\link{create_method}}
specification.}

\item{legend_pos}{Legend position.}

\item{legend_size}{Font size of legend.}

\item{...}{Pass generic \code{\link[base]{plot}} options.}
}
\value{
Plot object.
}
\description{
This function plots the non-negative ensemble learner weights for both either
a short-stacked or standard-stacked ensemble learner.
}
