% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{predict.ensemble}
\alias{predict.ensemble}
\title{Predict method for standard-stacked ensemble}
\usage{
\method{predict}{ensemble}(
  object,
  methods,
  X,
  Y,
  Xnew,
  quiet = TRUE,
  pb = NULL,
  pb_cf = NULL,
  pb_np = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"ensemble"} returned by \code{\link{ensemble}}.}

\item{methods}{List of methods created with \code{\link{create_method}}.}

\item{X}{Covariate matrix of the training sample.}

\item{Y}{Vector of outcomes of the training sample.}

\item{Xnew}{Covariate matrix of the test sample.}

\item{quiet}{Logical. If \code{FALSE}, progress output is printed to the console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_cf}{Integer indicating the current cross-fitting fold 
(for progress bar updates).}

\item{pb_np}{String indicating the current nuisance parameter being 
estimated or predicted (for progress bar updates).}

\item{...}{Ignored additional arguments.}
}
\value{
A list containing:
\describe{
  \item{cf_preds}{A matrix of predictions for the test sample from the individual learners.}
  \item{np}{A numeric vector containing aggregated ensemble predictions.}
}
}
\description{
Generate predictions from a trained ensemble standard-stacked learner 
created with \code{\link{ensemble}}.
}
\keyword{internal}
