% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.forest_drf_fit}
\alias{predict.forest_drf_fit}
\title{Predictions based on Distributional Random Forest}
\usage{
\method{predict}{forest_drf_fit}(forest_drf_fit, Xnew = NULL, functional = "mean", ...)
}
\arguments{
\item{forest_drf_fit}{Output of \code{\link{forest_drf_fit}}}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Prediction of fitted values (for a new set of covariates)
from a Random Forest.
}
\keyword{internal}
