% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.forest_grf_fit}
\alias{predict.forest_grf_fit}
\title{Predictions based on Random Forest}
\usage{
\method{predict}{forest_grf_fit}(forest_grf_fit, xnew = NULL, ...)
}
\arguments{
\item{xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}

\item{plasso_fit}{Output of \code{\link{forest_grf_fit}}}
}
\value{
Vector of fitted values.
}
\description{
Prediction of fitted values (for a potentially new set of covaraites xnew)
from a Random Forest.
}
\keyword{internal}
