% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.glm_fit}
\alias{predict.glm_fit}
\title{Predict method for binary GLM fits}
\usage{
\method{predict}{glm_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{glm_fit}}.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}

\item{...}{Ignored additional arguments.}
}
\value{
A numeric vector of probabilities for the positive class.
}
\description{
Generates predicted class probabilities from fitted logistic or probit regression models.
}
\keyword{internal}
