% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.knn_fit}
\alias{predict.knn_fit}
\title{Predictions based on the k-Nearest-Neighbor algorithm}
\usage{
\method{predict}{knn_fit}(args, x, y, xnew = NULL, weights = FALSE)
}
\arguments{
\item{args}{Output of \code{\link{knn_fit}}}

\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}

\item{xnew}{Covariate matrix of test sample}

\item{weights}{Always FALSE as}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for xnew}
\item{weights}{Binary matrix of dimension nrow(xnew) X nrow(x) that indicates
the nearest neighbors for a given test observation}
}
\description{
Predictions based on fitted the k-Nearest-Neighbor algorithm. Note that
if no \code{k} is explicitly specified, this evaluates to \code{k = 10} as
default value.
The method also provides prediction weights if required.
}
\keyword{internal}
