% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.knn_prop_fit}
\alias{predict.knn_prop_fit}
\title{Predict method for k-Nearest Neighbor fits}
\usage{
\method{predict}{knn_prop_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{knn_prop_fit}}.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}

\item{...}{Ignored additional arguments.}
}
\value{
A numeric vector of predicted probabilities for the positive class
  (binary outcome), or a matrix with one column per class (multiclass outcome).
}
\description{
Generates predicted class probabilities from fitted k-Nearest Neighbor models.
}
\keyword{internal}
