% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.lasso_fit}
\alias{predict.lasso_fit}
\title{Lasso prediction}
\usage{
\method{predict}{lasso_fit}(lasso_fit, Xnew = NULL, ...)
}
\arguments{
\item{lasso_fit}{Output of \code{\link{lasso_fit}}}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Prediction of fitted values based on fine-tuned Lasso regression model.
}
\keyword{internal}
