% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.logit_nnet_fit}
\alias{predict.logit_nnet_fit}
\title{Predict method for \code{nnet} logistic regression fits}
\usage{
\method{predict}{logit_nnet_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{logit_nnet_fit}}.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}

\item{...}{Ignored additional arguments.}
}
\value{
A numeric vector of predicted probabilities for the positive class
  (binary outcome), or a matrix with one column per class (multiclass outcome).
}
\description{
Generates predicted class probabilities from fitted logistic regression models.
}
\keyword{internal}
