% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.mean_fit}
\alias{predict.mean_fit}
\title{Predict after arithmetic mean fitting}
\usage{
\method{predict}{mean_fit}(mean_fit, xnew = NULL, ...)
}
\arguments{
\item{mean_fit}{Output of \code{\link{mean_fit}}}

\item{xnew}{Covariate matrix of test sample. If no test sample provided,
prediction is done for training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Predicts arithmetic mean.
}
\keyword{internal}
