% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.nb_bernoulli_fit}
\alias{predict.nb_bernoulli_fit}
\title{Predict method for Bernoulli Naive Bayes fits}
\usage{
\method{predict}{nb_bernoulli_fit}(nb_bernoulli_fit, X, Y, Xnew = NULL)
}
\arguments{
\item{nb_bernoulli_fit}{Output of \code{\link{nb_bernoulli_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}
}
\value{
Predicted class probabilities.  
  \describe{  
    \item{Binary outcome}{A numeric vector of probabilities for the positive class.}  
    \item{Multiclass outcome}{A numeric matrix with one column per class, containing the predicted probabilities.}  
  }
}
\description{
Generates predicted class probabilities from fitted Bernoulli Naive Bayes models.
}
\keyword{internal}
