% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.ols_fit}
\alias{predict.ols_fit}
\title{Predictions based on OLS}
\usage{
\method{predict}{ols_fit}(ols_fit, Xnew = NULL, ...)
}
\arguments{
\item{ols_fit}{Output of \code{\link{ols_fit}}}

\item{Xnew}{Covariate matrix of test sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Prediction based on fitted ordinary least squares models. The method also
provides prediction weights if required.
}
\keyword{internal}
