% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{predict.ovo_fit}
\alias{predict.ovo_fit}
\title{Predicts class probabilities using One-Vs-One (OvO) fitted classifiers}
\usage{
\method{predict}{ovo_fit}(ovo_fit, X, Y, Xnew = NULL, weights = FALSE, method = "logit")
}
\arguments{
\item{ovo_fit}{Output list of \code{\link{ovo_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{weights}{Flag indicating whether weights are used (unsupported for propensity score estimation)}

\item{method}{Method used for fitting the binary classifiers (e.g., "logit" or "xgboost")}
}
\value{
Returns list containing:
\item{prediction}{matrix of predictions for Xnew}
\item{weights}{Not supported for propensity score estimation}
}
\description{
Predicts class probabilities using One-Vs-One (OvO) fitted classifiers
}
\keyword{internal}
