% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_OvO_OvR.R
\name{predict.ovr_fit}
\alias{predict.ovr_fit}
\title{Predict method for One-vs-Rest (OvR) classifiers}
\usage{
\method{predict}{ovr_fit}(ovr_fit, X, Y, Xnew = NULL, method)
}
\arguments{
\item{ovr_fit}{Output list from \code{\link{ovr_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}

\item{method}{Character string specifying the binary classifier used for fitting.}
}
\value{
A data frame of predicted class probabilities with one column per class.  
  Probabilities are normalized to sum to one for each observation.
}
\description{
Generates predicted class probabilities using a set of fitted OvR binary classifiers.
}
\keyword{internal}
