% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.plasso_fit}
\alias{predict.plasso_fit}
\title{Predictions based on Post-Lasso regression}
\usage{
\method{predict}{plasso_fit}(plasso_fit, Xnew = NULL, ...)
}
\arguments{
\item{plasso_fit}{Output of \code{\link{plasso_fit}}}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Prediction of fitted values (for a potentially new set of covaraites Xnew)
based on a trained Post-Lasso model.
}
\keyword{internal}
