% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.ranger_fit}
\alias{predict.ranger_fit}
\title{Predict method for Probability Forest fits}
\usage{
\method{predict}{ranger_fit}(ranger_fit, X, Y, Xnew = NULL)
}
\arguments{
\item{ranger_fit}{Output of \code{\link{ranger_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}
}
\value{
A numeric matrix with one column per class, containing the predicted probabilities.
}
\description{
Generates predicted class probabilities from fitted Probability Forest models.
}
\keyword{internal}
