% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.ridge_fit}
\alias{predict.ridge_fit}
\title{Predict method for ridge regression models}
\usage{
\method{predict}{ridge_fit}(object, Xnew, ...)
}
\arguments{
\item{object}{An object of class \code{ridge_fit} from \code{\link{ridge_fit}}.}

\item{Xnew}{Numeric matrix of covariates for test sample.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of predictions for \code{Xnew}.
}
\description{
Generates predictions from a fitted ridge regression model.
Applies the same standardization used during training to new data.
}
\keyword{internal}
