% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.rlasso_fit}
\alias{predict.rlasso_fit}
\title{Predict method for hdm rlasso models}
\usage{
\method{predict}{rlasso_fit}(rlasso_fit, Xnew = NULL, ...)
}
\arguments{
\item{rlasso_fit}{An object of class \code{rlasso_fit} from \code{\link{rlasso_fit}}.}

\item{Xnew}{Numeric matrix of covariates for test sample. If \code{NULL},
predictions are generated for the training data.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of predicted values.
}
\description{
Generates predictions from a fitted post-lasso regression model
estimated using the \code{hdm} package.
}
\keyword{internal}
