% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.xgboost_prop_fit}
\alias{predict.xgboost_prop_fit}
\title{Predict method for \code{xgboost} fits}
\usage{
\method{predict}{xgboost_prop_fit}(xgboost_fit, X, Y, Xnew = NULL)
}
\arguments{
\item{xgboost_fit}{Output of \code{\link{xgboost_prop_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}
}
\value{
A data frame of predicted class probabilities.  
  \describe{  
    \item{Binary outcome}{A two-column data frame with probabilities for each class.}  
    \item{Multiclass outcome}{A data frame with one column per class, containing the predicted probabilities.}  
  }
}
\description{
Generates predicted class probabilities from fitted gradient boosting models.
}
\keyword{internal}
