% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_w_mat}
\alias{prep_w_mat}
\title{One-hot encoding of treatment vector}
\usage{
prep_w_mat(w)
}
\arguments{
\item{w}{Treatment vector.
Provide as factor to control ordering of the treatments,
otherwise arranged in ascending order or alphabetically.}
}
\value{
Logical matrix of treatment indicators (n x T+1).
}
\description{
\code{\link{prep_w_mat}} creates a matrix of binary treatment indicators
(n x T+1 with T being the number of treatments)
where each column contains a binary indicator for one treatment (i.e. one-hot
encoding).
}
