% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{prob_forest_fit}
\alias{prob_forest_fit}
\title{Calculates Probability Forest fit using the \code{\link{grf}} package}
\usage{
prob_forest_fit(X, Y, arguments = list(num.trees = 1000))
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{vector of outcomes}

\item{arguments}{List of arguments passed to  \code{\link{probability_forest}}}
}
\value{
An object with S3 class "probability_forest"
}
\description{
Calculates Probability Forest fit using the \code{\link{grf}} package
}
\keyword{internal}
