% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_methods}
\alias{process_methods}
\title{Method creation for ensemble (internal version)}
\usage{
process_methods(methods, NuPa, K, M)
}
\arguments{
\item{methods}{Either:
\enumerate{
  \item{A list of methods (like in \code{create_method}) that will be used for all nuisance parameters, or}
  \item{A named list where each name corresponds to a nuisance parameter and contains a list of methods for that parameter}
}}

\item{NuPa}{Character vector of nuisance parameters to be estimated}

\item{K}{Number of unique treatment statuses}

\item{M}{Number of unique outcome values}
}
\value{
Nested list object organized by nuisance parameter that will be passed as input to \code{\link{ensemble}}
}
\description{
Creates the methods to be actually used in the subsequent \code{\link{ensemble}} model,
allowing for either a simple method list or nuisance-parameter-specific method lists.
}
\keyword{internal}
