% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{rlasso_fit}
\alias{rlasso_fit}
\title{Fits Lasso regression using hdm::rlasso}
\usage{
rlasso_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{Y}{vector of outcomes}

\item{arguments}{List of arguments passed to \code{\link[hdm]{rlasso}}}
}
\value{
An object with S3 class "rlasso"
}
\description{
\code{\link{rlasso_fit}} estimates Lasso regression using the hdm package.
}
\keyword{internal}
