% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{svm_fit}
\alias{svm_fit}
\title{Fit Support Vector Machine using \code{e1071}}
\usage{
svm_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{arguments}{List of arguments passed to \code{e1071::svm()}.}
}
\value{
An object of class \code{svm} containing the fitted model.
}
\description{
Fits a Support Vector Machine (SVM) model using the \code{e1071} package.
Uses a radial kernel and \code{C-classification} by default, with probability estimates enabled.
}
\keyword{internal}
