% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tune_learners}
\alias{tune_learners}
\title{Tune learner hyperparameters}
\usage{
tune_learners(type = c("tune_full_sample", "tune_on_fold"), methods, X, Y)
}
\arguments{
\item{type}{Tuning type, either "\code{tune_full_sample}" or "\code{tune_on_fold}".}

\item{methods}{List of methods for ensemble estimation.}

\item{X}{Covariate matrix.}

\item{Y}{Numeric vector containing the outcome variable.}
}
\value{
Modified methods list with tuned parameters.
}
\description{
Automatically tunes hyperparameters for specified learners in a methods list
}
\keyword{internal}
