% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_progress}
\alias{update_progress}
\title{Update progress bar}
\usage{
update_progress(pb, pb_np, pb_cf, pb_cv, task, method)
}
\arguments{
\item{pb}{Progress bar object from `progress` package}

\item{pb_np}{Current nuisance parameter (character)}

\item{pb_cf}{Current cross-fitting fold number (integer)}

\item{pb_cv}{Current cross-validation fold number (integer)}

\item{task}{Task description (character)}

\item{method}{Method name (character)}
}
\description{
This function updates progress bar with standardized formatting
}
\keyword{internal}
