% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{weights.ensemble}
\alias{weights.ensemble}
\title{Smoother weights from cross-validated ensemble learner}
\usage{
\method{weights}{ensemble}(object, methods, X, Y, Xnew, ens_w = NULL, quiet = TRUE, ...)
}
\arguments{
\item{object}{Trained ensemble object from \code{\link{ensemble}}.}

\item{methods}{List of methods built via \code{\link{create_method}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{Xnew}{Covariate matrix of test sample.}

\item{ens_w}{Ensemble weights to aggregate predictions from different learners. 
Must be a vector of length \code{ncol(object$cf_preds)}. Optional.}

\item{quiet}{If \code{FALSE}, the method currently computed is printed to the console.}

\item{...}{Ignore unused arguments.}
}
\value{
A matrix of dimension \code{nrow(Xnew)} × \code{nrow(X)} containing ensemble smoother weights.
}
\description{
Extract ensemble smoother weights for a covariate matrix based on a
cross-validated ensemble model created by \code{\link{ensemble}}.
}
\keyword{internal}
