% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.ensemble_core}
\alias{weights.ensemble_core}
\title{Ensemble smoother weights}
\usage{
\method{weights}{ensemble_core}(object, method, X_tr, Y_tr, X_te, quiet = TRUE, ...)
}
\arguments{
\item{object}{List of fitted methods (from \code{\link{ensemble_core}}}

\item{method}{List of raw methods built via \code{\link{create_method}}}

\item{X_tr}{Covariate matrix of training sample}

\item{Y_tr}{Vector of outcomes of training sample}

\item{X_te}{Covariate matrix of test sample}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}

\item{...}{Ignore unused arguments}
}
\value{
Returns a three-dimensional array of dimension
\code{nrow(Xnew)} X \code{nrow(X)} X \code{length(method)}
containing the smoother weights that deliver predictions of each method
where each row gives the weight that each training outcome received in the
prediction of the respective test set observation.
}
\description{
Extract smoother weights for some (new) covariate matrix based on the fully
trained models of the ensemble model from \code{\link{ensemble_core}}.
}
