% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.ensemble_short}
\alias{weights.ensemble_short}
\title{Smoother weights from short-stacked ensemble learner}
\usage{
\method{weights}{ensemble_short}(
  object,
  methods,
  X,
  Y,
  cf_mat,
  subset = NULL,
  ens_w = NULL,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{object}{Short-stacked ensemble learner from \code{\link{ensemble_short}}.}

\item{methods}{List of method models from \code{\link{create_method}} used 
as input for \code{\link{ensemble_short}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{cf_mat}{Logical matrix with \code{k} columns of indicators representing 
folds (for example created by \code{\link{prep_cf_mat}}).}

\item{subset}{Logical vector indicating which observations to use for 
determining ensemble weights. If not provided, all observations are used.}

\item{ens_w}{Ensemble weights to aggregate predictions from different learners. 
Must be a vector of length \code{ncol(object$cf_preds)}. Optional.}

\item{quiet}{If \code{FALSE}, the method currently computed is printed to the console.}

\item{...}{Ignore unused arguments.}
}
\value{
A matrix of dimension \code{nrow(X)} × \code{nrow(X)} containing ensemble smoother weights.
}
\description{
Extract smoother weights from a short-stacked ensemble model created by
\code{\link{ensemble_short}}.
}
\keyword{internal}
