% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{weights.forest_grf_fit}
\alias{weights.forest_grf_fit}
\title{Smoother weights from Random Forest model}
\usage{
\method{weights}{forest_grf_fit}(forest_grf_fit, xnew, ...)
}
\arguments{
\item{xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}

\item{ridge_fit}{Output of \code{\link{forest_grf_fit}}}
}
\value{
Matrix of smoother weights.
}
\description{
Extract smoother weights for test sample from a Random Forest model.
}
\keyword{internal}
