% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.knn_fit}
\alias{weights.knn_fit}
\title{Smoother weights from k-nearest neighbor algorithm}
\usage{
\method{weights}{knn_fit}(arguments, X, Xnew = NULL, ...)
}
\arguments{
\item{arguments}{Output of \code{\link{knn_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Xnew}{Covariate matrix of test sample.}

\item{...}{Ignore unused arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Extract smoother weights from a k-nearest neighbor algorithm.
Each weight equals \code{1 / k} for neighbors and 0 for non-neighbors.
}
\keyword{internal}
