% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R, R/ml_wrapper.R
\name{weights.mean_fit}
\alias{weights.mean_fit}
\title{Arithmetic mean smoother weights}
\usage{
\method{weights}{mean_fit}(mean_fit, xnew, ...)

\method{weights}{mean_fit}(mean_fit, xnew, ...)
}
\arguments{
\item{mean_fit}{Output of \code{\link{mean_fit}}}

\item{xnew}{Covariate matrix of test sample}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of smoother weights.

Matrix of smoother weights.
}
\description{
Returns smoother weights for test sample based on
arithmetic mean fitting of the training sample.

Returns smoother weights for test sample based on
arithmetic mean fitting of the training sample.
}
\keyword{internal}
