% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{weights.ols_fit}
\alias{weights.ols_fit}
\title{Compute smoother weights from OLS}
\usage{
\method{weights}{ols_fit}(object, X, Xnew, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{ols_fit}}.}

\item{X}{Covariate matrix of the training sample.}

\item{Xnew}{Covariate matrix of the test sample.}

\item{...}{Ignored arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Generate smoother weights for a test sample from an OLS model.
}
\keyword{internal}
