% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.plasso_fit}
\alias{weights.plasso_fit}
\title{Compute smoother weights from post-Lasso}
\usage{
\method{weights}{plasso_fit}(plasso_fit, Xnew = NULL, ...)
}
\arguments{
\item{plasso_fit}{An object returned by \code{\link{plasso_fit}}.}

\item{Xnew}{Optional covariate matrix of the test sample. 
If not provided, predictions are computed for the training sample.}

\item{...}{Ignored arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Generate smoother weights for a test sample from a post-Lasso regression model.
}
\keyword{internal}
