% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R, R/ml_wrapper.R
\name{weights.ridge_fit}
\alias{weights.ridge_fit}
\title{Smoother weights from Ridge regression prediction}
\usage{
\method{weights}{ridge_fit}(ridge_fit, X, Y, Xnew = NULL, ...)

\method{weights}{ridge_fit}(ridge_fit, X, Y, Xnew = NULL, ...)
}
\arguments{
\item{ridge_fit}{Output of \code{\link{ridge_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of smoother weights.

Matrix of smoother weights.
}
\description{
Extract smoother weights for test sample from a Ridge regression model.

Extract smoother weights for test sample from a Ridge regression model.
}
\keyword{internal}
