% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.xgboost_fit}
\alias{weights.xgboost_fit}
\title{Smoother weights from XGBoost model}
\usage{
\method{weights}{xgboost_fit}(xgboost_fit, X, Y, Xnew = NULL, ...)
}
\arguments{
\item{xgboost_fit}{Output of \code{\link{xgboost_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Xnew}{Covariate matrix of test sample. 
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Extract smoother weights for test sample from an XGBoost model.
}
\keyword{internal}
