% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_stacking}
\alias{which_stacking}
\title{Short- or Standard-stacking message}
\usage{
which_stacking(cv = 1)
}
\arguments{
\item{cv}{Cross-validation value for ensemble weights.}
}
\value{
Message in console. No object is returned.
}
\description{
This function prints a message into the console whether short- or standard-
stacking is used.
}
\keyword{internal}
