% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{xgboost_fit}
\alias{xgboost_fit}
\title{Fits XGBoost}
\usage{
xgboost_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{vector of outcomes}

\item{arguments}{List of arguments passed to \code{\link[xgboost]{xgboost}}}
}
\value{
An object with S3 class \code{\link[xgboost]{xgboost}}
}
\description{
\code{\link{xgboost_fit}} fits an XGBoost using the
\code{\link{xgboost}} package.
}
\keyword{internal}
