% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{xgboost_prop_fit}
\alias{xgboost_prop_fit}
\title{Fit gradient boosting model using \code{xgboost}}
\usage{
xgboost_prop_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{arguments}{List of arguments passed to \code{xgboost::xgboost()}.}
}
\value{
An object of class \code{xgb.Booster} containing the fitted model.
}
\description{
Fits a gradient boosting model using the \code{xgboost} package.  
Uses \code{objective = "binary:logistic"} for binary outcomes and  
\code{objective = "multi:softprob"} for multiclass outcomes.
}
\keyword{internal}
