% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cluster_compatibility}
\alias{check_cluster_compatibility}
\title{Compatibility check of cluster vector}
\usage{
check_cluster_compatibility(cl, cf)
}
\arguments{
\item{cl}{A vector representing the cluster assignments}

\item{cf}{The desired number of folds for cross-validation}
}
\value{
Does not return anything, only throws an error if checks fail.
}
\description{
\code{\link{check_cluster_compatibility}} checks if the cross-fitting procedure is feasible
given the cluster vector and desired number of cross-fitting folds.
}
