% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{create_method}
\alias{create_method}
\title{Method creation for ensemble}
\usage{
create_method(
  method = c("mean", "ols", "ridge", "plasso", "forest_grf", "lasso", "knn",
    "forest_drf"),
  x_select = NULL,
  arguments = list(),
  name = NULL
)
}
\arguments{
\item{method}{Choose method from \code{c("mean", "ols", "ridge", "plasso", "forest_grf", "lasso", "knn", "forest_drf")}.
To be continued.}

\item{x_select}{Optional logical vector of length equal to the number of
columns of the covariate matrix indicating which variables should be used by
this method. E.g. tree-based methods usually should not be provided with the
interactions that Lasso is using.}

\item{arguments}{Optional list containing the additional arguments that should be
passed to the underlying method.}

\item{name}{Optional string naming the method.}
}
\value{
List object that can be passed as input to \code{\link{ensemble}}
}
\description{
Creates the methods to be used in the subsequent \code{\link{ensemble}} model.
}
\examples{
# create list of ml methods for ensemble
ml = list(
 "ols" = create_method("ols"),
 "forest_grf" = create_method("forest_grf"),
 "knn" = create_method("knn", arguments = list("k" = 3))
)

}
