% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cleandata.R
\name{data_screen}
\alias{data_screen}
\title{Data screening}
\usage{
data_screen(data, treat = NULL, bin_cut = 0.01, corr_cut = 0.99, print = FALSE)
}
\arguments{
\item{data}{Matrix the variables to be screened}

\item{treat}{Optional binary treatment vector if screening should be done within treatment groups}

\item{bin_cut}{Cut-off fraction under which nearly empty binary variables should be removed. Default 0.01.}

\item{corr_cut}{Cut-off above which highly correlated variables should be removed. Default 0.99.}

\item{print}{Shows details about the removed variables at each step}
}
\value{
Screened matrix.
}
\description{
\code{\link{data_screen}} takes a matrix of data and cleans it for the sake
of an improved input for subsequent prediction algorithms. \cr
It removes \cr
1. Variables without variation \cr
2. Dummy variables where one group is nearly empty (optionally in one of both treatment groups) \cr
3. Redundant (i.e. highly correlated) variables \cr
}
