% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cleandata.R
\name{design_matrix}
\alias{design_matrix}
\title{Creates design matrix}
\usage{
design_matrix(data, int = NULL, int_d = 2, poly = NULL, poly_d = 2, log = NULL)
}
\arguments{
\item{data}{Matrix with the main variables}

\item{int}{Vector of strings with the variables to be interacted}

\item{int_d}{Degree of interactions created. Default is 2.}

\item{poly}{Vector of strings with the variables for which polynomials should be created}

\item{poly_d}{Degree of polynomials to be created. Default is 2.}

\item{log}{Vector of strings with the variables for which the logged versions should be added}
}
\value{
Matrix including the main variables and the newly generated ones.
}
\description{
\code{\link{design_matrix}} creates interactions and polynomials for a proper
design matrix that can be used as input for prediction functions.
}
