% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble learner training}
\usage{
ensemble(ml, x, y, subset = NULL, nfolds = 5, quiet = TRUE)
}
\arguments{
\item{ml}{List of methods built via \code{\link{create_method}} to be used in
ensemble model}

\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}

\item{subset}{Logical vector indicating which observations to use for determining
ensemble weights. If not provided, all observations are used.}

\item{nfolds}{Number of folds used in cross-validation of ensemble weights
(default \code{nfolds = 5})}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}
}
\value{
List object containing:
\item{fit_cv}{matrix of dimension \code{nrow(x)} x \code{length(ml)} containing
the cross-fitted predictions of the machine learning methods from the ensemble}
\item{nnls_weights}{the weights that each machine learning method receives in
the ensemble}
\item{ml}{list of fitted machine learning models from all cross-fitting folds
(if storeModels is not set to "No")}
}
\description{
\code{\link{ensemble}} trains an ensemble learner on the given training data.
}
