% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Train ensemble learner with standard stacking}
\usage{
ensemble(
  methods,
  X,
  Y,
  nfolds = 5,
  quiet = TRUE,
  pb = NULL,
  pb_np = NULL,
  pb_cf = NULL,
  pb_cv = NULL
)
}
\arguments{
\item{methods}{List of methods created with \code{\link{create_method}} to be 
used in the ensemble model.}

\item{X}{Covariate matrix of the training sample.}

\item{Y}{Vector of outcomes of the training sample.}

\item{nfolds}{Integer. Number of folds used in cross-validation of 
ensemble weights. Default is 5.}

\item{quiet}{Logical. If \code{FALSE}, progress output is printed to the console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_np}{String indicating the current nuisance parameter being 
estimated or predicted (for progress bar updates).}

\item{pb_cf}{Integer indicating the current cross-fitting fold 
(for progress bar updates).}

\item{pb_cv}{Integer indicating the current cross-validation fold 
(for progress bar updates).}
}
\value{
A list with class \code{"ensemble"} containing:
\describe{
  \item{cf_preds}{A matrix of dimension \code{nrow(X)} x \code{length(methods)} 
   containing the cross-fitted predictions of the machine learning methods from the ensemble.}
  \item{nnls_w}{A numeric vector of weights that each machine learning 
   method receives in the ensemble.}
  \item{ens_models}{A list of fitted machine learning models from the full sample.}
}
}
\description{
Train an ensemble learner on the given training data using cross-validation for ensemble weights.
}
\keyword{internal}
