% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble learner training}
\usage{
ensemble(
  method,
  X,
  Y,
  nfolds = 5,
  cv_mat = NULL,
  quiet = TRUE,
  pb = NULL,
  pb_np = NULL,
  pb_cf = NULL,
  pb_cv = NULL
)
}
\arguments{
\item{method}{List of methods built via \code{\link{create_method}} to be used in
ensemble model}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{nfolds}{Number of folds used in cross-validation of ensemble weights
(default \code{nfolds = 5})}

\item{cv_mat}{Optional list of logical matrices of indicators representing the different folds,
possibly from already estimated \code{NuisanceParameters} object.}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_np}{String indicating the current nuisance parameter being
estimated/predicted (for progress bar updates).}

\item{pb_cf}{Integer indicating the current cross-fitting fold
(for progress bar updates).}

\item{pb_cv}{Integer indicating the current cross-validation fold
(for progress bar updates).}
}
\value{
List object containing:
\item{fit_cv}{matrix of dimension \code{nrow(X)} X \code{length(method)} containing
the cross-fitted predictions of the machine learning methods from the ensemble}
\item{nnls_weights}{the weights that each machine learning method receives in
the ensemble}
\item{method}{list of fitted machine learning models from all cross-fitting folds
(if storeModels is not set to "No")}
}
\description{
\code{\link{ensemble}} trains an ensemble learner on the given training data.
}
