% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_core}
\alias{ensemble_core}
\title{Ensemble fitting}
\usage{
ensemble_core(ml, x_tr, y_tr, quiet = TRUE)
}
\arguments{
\item{ml}{List of methods built via \code{\link{create_method}} to be used in
ensemble model}

\item{x_tr}{Covariate matrix of training sample}

\item{y_tr}{Vector of outcomes of training sample}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}
}
\value{
Returns list object containing the trained models of the ensemble.
}
\description{
Core function of \code{\link{ensemble}} used for fitting (i.e. training) the
inputted (machine learning) methods for the ensemble model.
}
\keyword{internal}
