% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_core}
\alias{ensemble_core}
\title{Ensemble core fitting}
\usage{
ensemble_core(
  ml,
  X_tr,
  Y_tr,
  quiet = TRUE,
  pb = NULL,
  pb_cf = NULL,
  pb_cv = NULL,
  pb_np = NULL
)
}
\arguments{
\item{ml}{List of methods built via \code{\link{create_method}} to be used in
ensemble model}

\item{X_tr}{Covariate matrix of training sample}

\item{Y_tr}{Vector of outcomes of training sample}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}
}
\value{
Returns list object containing the trained models of the ensemble.
}
\description{
Core function of \code{\link{ensemble}} used for fitting (i.e. training) the
inputted (machine learning) methods for the ensemble model for one given set
of covariates and targets.
}
