% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_core}
\alias{ensemble_core}
\title{Ensemble core fitting}
\usage{
ensemble_core(
  method,
  X_tr,
  Y_tr,
  quiet = TRUE,
  pb = NULL,
  pb_cf = NULL,
  pb_cv = NULL,
  pb_np = NULL
)
}
\arguments{
\item{method}{List of methods built via \code{\link{create_method}} to be used in
ensemble model.}

\item{X_tr}{Covariate matrix of training sample.}

\item{Y_tr}{Vector of outcomes of training sample.}

\item{quiet}{If FALSE, progress output is printed into console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_cf}{Integer indicating the current cross-fitting fold
(for progress bar updates).}

\item{pb_cv}{Integer indicating the current cross-validation fold
(for progress bar updates).}

\item{pb_np}{String indicating the current nuisance parameter being
estimated/predicted (for progress bar updates).}
}
\value{
Returns list object containing the trained models of the ensemble.
}
\description{
Core function of \code{\link{ensemble}} used for fitting (i.e. training) the
inputted (machine learning) methods for the ensemble model for one given set
of covariates and targets.
}
\keyword{internal}
