% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_short}
\alias{ensemble_short}
\title{Train ensemble learner with short stacking}
\usage{
ensemble_short(
  methods,
  X,
  Y,
  cf_mat,
  subset = NULL,
  storeModels = c("No", "Memory", "Disk"),
  quiet = TRUE,
  pb = NULL,
  pb_np = NULL
)
}
\arguments{
\item{methods}{List of methods created with \code{\link{create_method}} to be 
used in the ensemble model.}

\item{X}{Covariate matrix of the training sample.}

\item{Y}{Vector of outcomes of the training sample.}

\item{cf_mat}{Logical matrix with \code{k} columns of indicators representing 
the different folds.}

\item{subset}{Logical vector indicating which observations to use for 
determining ensemble weights. If not provided, all observations are used.}

\item{storeModels}{Character string specifying whether to store the models. 
Must be one of \code{c("No", "Memory", "Disk")} with "No" as the default.}

\item{quiet}{Logical. If \code{FALSE}, progress output is printed to the console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_np}{String indicating the current nuisance parameter being 
estimated or predicted (for progress bar updates).}
}
\value{
A list with class \code{"ensemble_short"} containing:
\describe{
  \item{cf_preds}{A matrix of dimension \code{nrow(X)} x \code{length(methods)} 
   containing the cross-fitted predictions of the machine learning methods from the ensemble.}
  \item{ens_models}{A list of fitted machine learning models from all 
   cross-fitting folds (if \code{storeModels} is not set to "No").}
}
}
\description{
Train an ensemble learner on the given training data using a short-stacking procedure.
}
\keyword{internal}
