% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_short.R
\name{ensemble_short}
\alias{ensemble_short}
\title{Ensemble learner training for short stacking}
\usage{
ensemble_short(
  ml,
  x,
  y,
  cf_mat,
  storeModels = c("No", "Memory", "Disk"),
  quiet = TRUE
)
}
\arguments{
\item{ml}{List of methods built via \code{\link{create_method}} to be used in
ensemble model.}

\item{x}{Covariate matrix of training sample.}

\item{y}{Vector of outcomes of training sample.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{storeModels}{Specifies whether and where to store the models.
If smoother weights should be computed in a second step (using the weights method),
the models need to be stored somewhere (either in memory or on disk).
Value has to be an item from \code{c("No", "Memory", "Disk")} with "No" being
the default.}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}
}
\value{
List object containing:
\item{fit_cv}{matrix of dimension \code{nrow(x)} x \code{length(ml)} containing
the cross-fitted predictions of the machine learning methods from the ensemble}
\item{nnls_weights}{the weights that each machine learning method receives in
the ensemble}
\item{mse_cv}{cross-validated MSEs for each machine learning  method}
\item{ml}{list of fitted machine learning models from all cross-fitting folds
(if storeModels is not set to "No")}
}
\description{
\code{\link{ensemble_short}} trains an ensemble learner on the given training data
in the spirit of a short-stacking procedure.
}
