% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{forest_drf_fit}
\alias{forest_drf_fit}
\title{Fits Distributional Random Forest}
\usage{
forest_drf_fit(x, y, arguments = list())
}
\arguments{
\item{x}{Matrix of covariates}

\item{y}{vector of outcomes}

\item{arguments}{List of arguments passed to \code{\link[drf]{drf}}}
}
\value{
An object with S3 class \code{\link[drf]{drf}}
}
\description{
\code{\link{forest_drf_fit}} fits a distributional random forest using the
\code{\link{drf}} package.
}
\keyword{internal}
