% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{forest_grf_fit}
\alias{forest_grf_fit}
\title{Fit random forest using grf package}
\usage{
forest_grf_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{arguments}{List of arguments passed to \code{\link[grf]{regression_forest}}.}
}
\value{
An object of class \code{forest_grf_fit} containing the fitted
  \code{\link[grf]{regression_forest}} model.
}
\description{
Fits a random forest for regression using the \code{grf} package.
}
\keyword{internal}
