% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_weights}
\alias{format_weights}
\title{Format ensemble weights for both standard and cross-validated stacking}
\usage{
format_weights(ens_weights, cv = 1)
}
\arguments{
\item{ens_weights}{List of ensemble weights to be formatted}

\item{cv}{Integer indicating cross-validation folds}
}
\value{
A data.frame (cv=1) or list of data.frames (cv>1) with formatted weights
}
\description{
Format ensemble weights for both standard and cross-validated stacking
}
\keyword{internal}
