% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{get_outcome_weights}
\alias{get_outcome_weights}
\title{Extraction of smoother weights for ensemble learner}
\usage{
get_outcome_weights(np_object, NuPa, quiet = TRUE, subset = NULL, ...)
}
\arguments{
\item{np_object}{A \code{NuisanceParameters} object created by 
\code{\link{nuisance_parameters}} function}

\item{NuPa}{String vector specifying which nuisance parameters to extract smoother matrices for.}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}

\item{subset}{Logical vector indicating which observations to use for determining
ensemble weights. If not provided, all observations are used.}

\item{...}{Additional arguments}
}
\value{
Matrix of dimension \code{nrow(X)} X \code{nrow(X)} containing
ensemble smoother weights, or list of such matrices if multiple models are requested.
}
\description{
Extract smoother weights from short- or standard-stacked ensemble model 
created by \code{\link{ensemble_short}} or \code{\link{ensemble}}.
}
