% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother}
\alias{get_smoother}
\title{Smoother weights from ensemble learner}
\usage{
get_smoother(object, methods, X, Y, subset = NULL, cf_mat, cv, quiet = TRUE)
}
\arguments{
\item{object}{Stacked ensemble learner from \code{\link{ensemble_short}} or \code{\link{ensemble}}.}

\item{methods}{List of methods built via \code{\link{create_method}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{subset}{Logical vector indicating which observations to use when determining
ensemble weights. If not provided, all observations are used.}

\item{cf_mat}{Logical matrix with \code{k} columns of indicators representing folds,
typically created by \code{\link{prep_cf_mat}}.}

\item{cv}{Integer specifying the number of cross-validation folds used in estimating
the ensemble model.}

\item{quiet}{If \code{FALSE}, the method currently computed is printed to the console.}
}
\value{
A matrix of dimension \code{nrow(X)} × \code{nrow(X)} containing ensemble smoother weights.
}
\description{
Extract smoother weights from a short- or standard-stacked ensemble model
created by \code{\link{ensemble_short}} or \code{\link{ensemble}}.
}
\keyword{internal}
