% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{get_smoother}
\alias{get_smoother}
\title{Extraction of smoother weights for ensemble learner}
\usage{
get_smoother(object, method, X, Y, subset = NULL, cf_mat, cv, quiet = TRUE)
}
\arguments{
\item{object}{Stacked ensemble learner from \code{\link{ensemble_short}} or \code{\link{ensemble}}.}

\item{method}{List of models by \code{\link{create_method}} that was used as
input for \code{\link{ensemble_short}} or \code{\link{ensemble}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{subset}{Logical vector indicating which observations to use for determining
ensemble weights. If not provided, all observations are used.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{cv}{Number of cross-validation folds when estimating ensemble model.}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}
}
\value{
Matrix of dimension \code{nrow(X)} X \code{nrow(X)} containing
ensemble smoother weights.
}
\description{
Extract smoother weights from short- or standard-stacked ensemble model 
created by \code{\link{ensemble_short}} or \code{\link{ensemble}}.
}
