% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{kl_convergence}
\alias{kl_convergence}
\title{Computes KL divergence-based convergence measure}
\usage{
kl_convergence(p, q_matrix)
}
\arguments{
\item{p}{Probability vector representing a probability distribution}

\item{q_matrix}{Square matrix representing transition probabilities}
}
\value{
Returns the KL divergence-based convergence measure
}
\description{
Computes KL divergence-based convergence measure
}
