% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{knn_fit}
\alias{knn_fit}
\title{Pseudo fitting of k-Nearest-Neighbour to training data}
\usage{
knn_fit(x, y, args = list())
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{vector of outcomes}

\item{args}{List of arguments passed to \code{\link[FastKNN]{k.nearest.neighbors}}}
}
\value{
A list of possible arguments for \code{\link[FastKNN]{k.nearest.neighbors}}
}
\description{
\code{\link{knn_fit}} is a pseudo fitting function where only the kNN object
is initialized but not fitted as the fitting process comes naturally with
the prediction part.
}
\keyword{internal}
