% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{knn_fit}
\alias{knn_fit}
\title{Pseudo fitting of k-Nearest-Neighbour to training data}
\usage{
knn_fit(arguments = list(), ...)
}
\arguments{
\item{arguments}{arguments of arguments passed to \code{\link[FastKNN]{k.nearest.neighbors}}}

\item{...}{Ignore unused arguments}
}
\value{
A list of possible arguments for \code{\link[FastKNN]{k.nearest.neighbors}}.
}
\description{
\code{\link{knn_fit}} is a pseudo fitting function where only the kNN object
is initialized but not fitted as the fitting process comes naturally with
the prediction part.
}
\keyword{internal}
