% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{knn_prop_fit}
\alias{knn_prop_fit}
\title{Fit k-Nearest Neighbor using \code{kknn}}
\usage{
knn_prop_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{arguments}{List of arguments passed to \code{kknn::train.kknn()}.}
}
\value{
An object of class \code{kknn} containing the fitted model.
}
\description{
Fits a k-Nearest Neighbor (kNN) model using the \code{kknn} package.
}
\keyword{internal}
