% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{lasso_fit}
\alias{lasso_fit}
\title{Fit lasso regression model}
\usage{
lasso_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{arguments}{List of arguments passed to \code{\link[glmnet]{cv.glmnet}}.}
}
\value{
An object of class \code{lasso_fit} containing the fitted
  \code{\link[glmnet]{cv.glmnet}} model.
}
\description{
Estimates lasso regression using the \code{glmnet} package.
Selects optimal lambda penalty parameter via cross-validation.
}
\keyword{internal}
