% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{logit_fit}
\alias{logit_fit}
\title{Fit (multinomial) logistic regression using \code{glmnet}}
\usage{
logit_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{arguments}{List of arguments passed to \code{glmnet::glmnet()}}
}
\value{
An object of class \code{glmnet} containing the fitted model.
}
\description{
Uses \code{family = "binomial"} for binary outcomes and \code{family = "multinomial"}
for multiclass outcomes.
}
\keyword{internal}
