% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{logit_nnet_fit}
\alias{logit_nnet_fit}
\title{Fit (multinomial) logistic regression using \code{nnet}}
\usage{
logit_nnet_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{arguments}{List of arguments passed to \code{nnet::multinom()}.}
}
\value{
An object of class \code{nnet} containing the fitted model.
}
\description{
Fits a logistic regression model using the \code{nnet} package.
Uses \code{multinom()} for binary and multiclass outcomes.
}
\keyword{internal}
