% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{mean_fit}
\alias{mean_fit}
\title{Fit arithmetic mean model}
\usage{
mean_fit(X, Y, ...)
}
\arguments{
\item{X}{Covariate matrix (ignored, for compatibility).}

\item{Y}{Numeric vector of training outcomes.}

\item{...}{Ignored additional arguments.}
}
\value{
An object of class `mean_fit` containing:
  \item{mean}{The arithmetic mean of `Y`}
  \item{N}{Number of observations}
}
\description{
Calculates the arithmetic mean of the training outcomes for use as a constant prediction model.
}
\keyword{internal}
