% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{mean_fit}
\alias{mean_fit}
\title{Arithmetic mean}
\usage{
mean_fit(x, y)
}
\arguments{
\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}
}
\value{
Returns list containing mean and number of observations.
}
\description{
\code{\link{mean_fit}} acts as prediction function by calculating the
arithmetic mean of the training outcomes.
}
\keyword{internal}
