% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnls_weights}
\alias{nnls_weights}
\title{Non-negative least squares function for estimation of ensemble weights}
\usage{
nnls_weights(X, Y)
}
\arguments{
\item{X}{A matrix where each column represents a different predictor variable
and each row represents an observation}

\item{Y}{A numeric vector of actual target values}
}
\value{
A numeric vector of the non-negative least weights.
}
\description{
\code{\link{nnls_weights}} estimates ensemble weights for a prediction problem
on the basis of the non-negative least squares algorithm that puts a positive
constraint on the regression coefficients.
}
\examples{
X = matrix(c(0.2, 0.3, 0.4, 0.5, 0.6, 0.7), ncol = 2)
Y = c(0.25, 0.45, 0.65)
\donttest{nnls_w = nnls_weights(X, Y)}

}
