% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_parameters.R
\name{nuisance_cf}
\alias{nuisance_cf}
\title{Cross-fitting of nuisance parameters}
\usage{
nuisance_cf(
  ml,
  Y,
  X,
  cf_mat,
  learner = c("t", "s", "both"),
  cv = 5,
  subset = NULL,
  storeModels = c("No", "Memory", "Disk"),
  path,
  quiet = TRUE,
  pb = NULL,
  pb_np = NULL
)
}
\arguments{
\item{ml}{List of methods to be used in \code{\link{ensemble}} estimation.
Methods can be created by \code{\link{create_method}}.}

\item{Y}{Vector of variable to be predicted.}

\item{X}{Matrix of covariates.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{learner}{Vector of characters indicating whether to use S or T learner
or a combination of the two.}

\item{cv}{Number of cross-validation folds when estimating ensemble model.
Default value is 1 which then evaluates to a short-stacking procedure which
is computationally less demanding than standard stacking.}

\item{subset}{Optional logical vector if only subset of data should be used for prediction.}

\item{storeModels}{Vector of characters indicating where to save individual 
models for future processing (default "No").}

\item{path}{Path to save fit_cv matrix and non-negative least square weights to.
Optionally, weights as well.}

\item{quiet}{If FALSE, progress output is printed into console.}
}
\value{
Returns a vector of length N containing nuisance parameters.
}
\description{
\code{\link{nuisance_cf}} makes a cross-fitted ensemble prediction using
\code{\link{ensemble}}.
}
\keyword{internal}
