% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_parameters.R
\name{nuisance_cf}
\alias{nuisance_cf}
\title{Cross-fitting of nuisance parameters}
\usage{
nuisance_cf(
  ml,
  y,
  x,
  cf_mat,
  cv = 5,
  subset = NULL,
  weights = FALSE,
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{ml}{List of methods to be used in \code{\link{ensemble}} estimation.
Methods can be created by \code{\link{create_method}}.}

\item{y}{Vector of variable to be predicted}

\item{x}{Matrix of covariates}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}})}

\item{cv}{Number of cross-validation folds when estimating ensemble model.
Default value is 1 which then evaluates to a short-stacking procedure which
is computationally less demanding than standard stacking.}

\item{subset}{Optional logical vector if only subset of data should be used for prediction}

\item{weights}{If TRUE, prediction weights of the outcome nuisance extracted and saved (requires to provide a path)}

\item{path}{Optional path to save the smoother weights of the ensemble model to.
Only considered, if \code{weights = TRUE}.}

\item{quiet}{If FALSE, method that is currently running is printed into console.}
}
\value{
Returns a vector of length n containing nuisance parameters.
}
\description{
\code{\link{nuisance_cf}} makes a cross-fitted ensemble prediction using
\code{\link{ensemble}}.
}
\keyword{internal}
