% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_parameters.R
\name{nuisance_cf}
\alias{nuisance_cf}
\title{Cross-fitting of nuisance parameters}
\usage{
nuisance_cf(
  method,
  Y,
  X,
  cf_mat,
  cv = 5,
  subset = NULL,
  storeModels = c("No", "Memory", "Disk"),
  quiet = TRUE,
  pb = NULL,
  pb_np = NULL
)
}
\arguments{
\item{method}{List of methods to be used in \code{\link{ensemble}} estimation.}

\item{Y}{Vector of variable to be predicted.}

\item{X}{Matrix of covariates.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds.}

\item{cv}{Number of cross-validation folds when estimating ensemble model (default: 1).}

\item{subset}{Optional logical vector if only subset of data should be used for prediction.}

\item{storeModels}{Vector of characters indicating where to save individual models for future processing.}

\item{quiet}{If FALSE, progress output is printed into console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_np}{String indicating the current nuisance parameter being
estimated/predicted (for progress bar updates).}
}
\value{
A list containing:
\itemize{
  \item \code{np}: A vector (or matrix for multinomial outcomes) of length N 
                   containing nuisance parameter estimates
  \item \code{models}: An object of class \code{ens.learner} containing fitted ensemble models
  \item \code{ens_weights}: A list of stacking weights
 }
}
\description{
\code{\link{nuisance_cf}} makes a cross-fitted ensemble prediction using
\code{\link{ensemble}}.
}
\keyword{internal}
