% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_parameters.R
\name{nuisance_cf}
\alias{nuisance_cf}
\title{Cross-fitting of nuisance parameters}
\usage{
nuisance_cf(
  methods,
  Y,
  X,
  cf_mat,
  cv = 5,
  subset = NULL,
  storeModels = c("No", "Memory", "Disk"),
  tuneLearners,
  do_bfgs = FALSE,
  quiet = TRUE,
  pb = NULL,
  pb_np = NULL
)
}
\arguments{
\item{methods}{List of methods to be used in \code{\link{ensemble}} estimation.}

\item{Y}{Vector of variable to be predicted.}

\item{X}{Matrix of covariates.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds.}

\item{cv}{Number of cross-validation folds when estimating ensemble model (default: 1).}

\item{subset}{Optional logical vector if only subset of data should be used for prediction.}

\item{storeModels}{Vector of characters indicating where to save individual models for future processing.}

\item{tuneLearners}{Optional hyperparameter tuning for selected learners (see \code{\link{create_method}} 
                  for details). Either \code{NULL} (default) for no tuning, or one of:
\describe{
  \item{\code{"full_sample" }}{Tune hyperparameters using full sample.}
  \item{\code{"fold" }}{Tuning is performed on the estimation part of the 
                       cross-fitting split, which is roughly
                       \eqn{F} times more computationally demanding.}
}}

\item{do_bfgs}{Logical. If \code{TRUE} and the outcome is multinomial,
estimates ensemble weights by BFGS optimization with a softmax 
constraint. If \code{FALSE}, estimates weights via stacked NNLS 
from the \code{nnls} package.}

\item{quiet}{If FALSE, progress output is printed into console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_np}{String indicating the current nuisance parameter being
estimated/predicted (for progress bar updates).}
}
\value{
A list containing:
\itemize{
  \item \code{np}: A vector (or matrix for multinomial outcomes) of length N 
                   containing nuisance parameter estimates
  \item \code{models}: An object of class \code{ens.learner} containing fitted ensemble models
 }
}
\description{
\code{\link{nuisance_cf}} makes a cross-fitted ensemble prediction using
\code{\link{ensemble}}.
}
\keyword{internal}
