% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_parameters.R
\name{nuisance_parameters}
\alias{nuisance_parameters}
\title{Nuisance parameter estimation}
\usage{
nuisance_parameters(
  NuPa = c("Y.hat", "Yw.hat", "Yz.hat", "W.hat", "Wz.hat", "Z.hat"),
  ml,
  x,
  y = NULL,
  w = NULL,
  z = NULL,
  cf = 5,
  cl = NULL,
  cv = 5,
  learner = c("t", "s", "both"),
  storeModels = c("No", "Memory", "Disk"),
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{NuPa}{String vector specifying the nuisance parameters to be estimated.
Currently supported: \code{c("Y.hat","Yw.hat","Yz.hat","W.hat","Wz.hat","Z.hat")}}

\item{ml}{List of methods to be used in \code{\link{ensemble}} estimation.
Methods can be created by \code{\link{create_method}}.}

\item{x}{Covariate matrix.}

\item{y}{Numerical vector containing the outcome variable.}

\item{w}{Numerical vector containing the treatment variable.}

\item{z}{Numerical vector containing the instrument variable.}

\item{cf}{Number of cross-fitting folds}

\item{cl}{Optional vector of cluster variable if cross-fitting should account
for clusters within the data.}

\item{cv}{Number of cross-validation folds when estimating ensemble (default 5).}

\item{learner}{Vector of characters indicating whether to use S or T learner
or a combination of the two.}

\item{storeModels}{Vector of characters indicating where to save individual 
models for future processing (default "No").}

\item{path}{Optional path to save the \code{\link{ensemble}} objects for later processing.
Saved as Ensemble_Yi where i is the number of the treatment in multiple treatment settings.}

\item{quiet}{If FALSE, progress output is printed into console.}
}
\value{
A list containing:
\itemize{
  \item{predictions:  }{Requested nuisance parameters}
  \item{models (optional):  }{\code{\link{ensemble}} objects (individual models)}
}
}
\description{
Nuisance parameter estimation
}
