% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_parameters.R
\name{nuisance_parameters}
\alias{nuisance_parameters}
\title{Nuisance parameter estimation}
\usage{
nuisance_parameters(
  NuPa = c("Y.hat", "Y.hat.d", "Y.hat.z", "D.hat", "D.hat.z", "Z.hat"),
  X,
  Y,
  D,
  Z,
  ml,
  cluster = NULL,
  stratify = FALSE,
  cf,
  stacking,
  learner = c("t", "s", "both"),
  storeModels = c("No", "Memory", "Disk"),
  path = NULL,
  quiet = TRUE
)
}
\arguments{
\item{NuPa}{String vector specifying the nuisance parameters to be estimated.
Currently supported: \code{c("Y.hat","Y.hat.d","Y.hat.z","D.hat","D.hat.z","Z.hat")}}

\item{X}{Covariate matrix.}

\item{Y}{Numerical vector containing the outcome variable.}

\item{D}{Numerical vector containing the treatment variable.}

\item{Z}{Numerical vector containing the instrument variable.}

\item{ml}{List of methods to be used in \code{\link{ensemble}} estimation.
Methods can be created by \code{\link{create_method}}.}

\item{learner}{Vector of characters indicating whether to use S or T learner
or a combination of the two.}

\item{storeModels}{Vector of characters indicating where to save individual 
models for future processing (default "No").}

\item{path}{Optional path to save the \code{\link{ensemble}} objects for later processing.
Saved as a list of models.}

\item{quiet}{If FALSE, progress output is printed into console.}

\item{d_mat}{Logical matrix of treatment indicators (N x T+1). For example created by \code{\link{prep_w_mat}}.}

\item{z_mat}{Logical matrix of instrument indicators (N x T+1). For example created by \code{\link{prep_w_mat}}.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{cl}{Optional vector of cluster variable if cross-fitting should account
for clusters within the data.}

\item{cv}{Number of cross-validation folds when estimating ensemble (default 5).}
}
\value{
A list containing:
\itemize{
  \item{predictions:  }{Requested nuisance parameters}
  \item{models (optional):  }{\code{\link{ensemble}} objects (individual models)}
}
}
\description{
Nuisance parameter estimation
}
