% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuisanceParameters-package.R
\docType{package}
\name{NuisanceParameters-package}
\alias{NuisanceParameters}
\alias{NuisanceParameters-package}
\title{NuisanceParameters: Nuisance Parameters Estimation}
\description{
Estimation of nuisance parameters using different prediction algorithms.
}
\author{
\strong{Maintainer}: Knaus Michael C. \email{michael.knaus@uni-tuebingen.de} (\href{https://orcid.org/0000-0002-7328-1363}{ORCID})

Other contributors:
\itemize{
  \item Glaisner Stefan \email{stefanglaisner@gmail.com} [contributor]
  \item Rakov Roman \email{roman.rakov.me@gmail.com} [contributor]
}

}
\keyword{internal}
