% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{ols_fit}
\alias{ols_fit}
\title{Fit ordinary least squares model}
\usage{
ols_fit(X, Y, ...)
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{...}{Ignored additional arguments.}
}
\value{
An object of class \code{ols_fit} containing estimated coefficients
        and the training covariate matrix.
}
\description{
Fits an ordinary least squares (OLS) regression to the given data using
\code{stats::lm.fit()}. An intercept is always added to the covariate matrix.
}
\keyword{internal}
