% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{ols_fit}
\alias{ols_fit}
\title{Fit ordinary least squares model}
\usage{
ols_fit(X, Y, ...)
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of OLS coefficients with class \code{ols_fit}.
}
\description{
Fits an ordinary least squares (OLS) regression to the given data using
\code{stats::lm.fit()}.
}
\keyword{internal}
