% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_OvO_OvR.R
\name{ovo_fit}
\alias{ovo_fit}
\title{Fits One-Vs-One (OvO) binary classifiers}
\usage{
ovo_fit(X, Y, method, parallel = FALSE, quiet = TRUE)
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{method}{Method to use for fitting binary classifiers (e.g., "logit" or "xgboost").}

\item{parallel}{Logical. Tries to run in parallel if \code{foreach} and 
\code{doParallel} are available. Defaults to FALSE.}

\item{quiet}{Logical. If FALSE, messages about parallel/sequential choice are printed.}
}
\value{
List of fitted OvO binary classifiers.
}
\description{
Fits One-Vs-One (OvO) binary classifiers for multi-class classification
with optional parallelization if \code{foreach} and \code{doParallel} packages are available.
}
\keyword{internal}
