% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{ovr_fit}
\alias{ovr_fit}
\title{Fits One-Vs-Rest (OvR) binary classifiers for multi-class classification}
\usage{
ovr_fit(X, Y, method = "logit")
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{method}{Method to use for fitting binary classifiers (e.g., "logit" or "xgboost")}
}
\value{
List of fitted OvR binary classifiers
}
\description{
Fits One-Vs-Rest (OvR) binary classifiers for multi-class classification
}
