% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{plasso_fit}
\alias{plasso_fit}
\title{Fits Post-Lasso regression}
\usage{
plasso_fit(x, y, arguments = list())
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{vector of outcomes}

\item{arguments}{List of arguments passed to \code{\link[plasso]{cv.plasso}}}
}
\value{
An object with S3 class \code{\link[plasso]{cv.plasso}}
}
\description{
\code{\link{plasso_fit}} estimates cross-validated Post-Lasso regression.
}
\keyword{internal}
