% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{predict.ensemble}
\alias{predict.ensemble}
\title{Prediction of fitted values by cross-validated ensemble learner}
\usage{
\method{predict}{ensemble}(
  object,
  ml,
  X,
  Y,
  Xnew,
  quiet = TRUE,
  pb = NULL,
  pb_cf = NULL,
  pb_np = NULL,
  ...
)
}
\arguments{
\item{object}{Trained ensemble object from \code{\link{ensemble}}}

\item{ml}{List of raw methods built via \code{\link{create_method}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}

\item{...}{Ignore unused arguments}
}
\value{
Vector containing ensemble predictions.
}
\description{
\code{\link{predict.ensemble}} makes prediction of fitted values for a cross-
validated ensemble learner.
}
