% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{predict.ensemble}
\alias{predict.ensemble}
\title{Ensemble learner prediction}
\usage{
\method{predict}{ensemble}(object, x, y, xnew, weights = FALSE, quiet = TRUE)
}
\arguments{
\item{object}{Trained ensemble object from \code{\link{ensemble}}}

\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}

\item{xnew}{Covariate matrix of test sample}

\item{weights}{If TRUE, the weights underlying the ensemble prediction for
xnew are calculated}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}
}
\value{
Predictions on the basis of ensemble object containing:
\item{p_ensemble}{the predictions of the trained ensemble for the test sample}
\item{p_best}{the predictions of the best performing single method for the test sample}
\item{p_all}{matrix containing predictions of all considered methods}
\item{w}{If \code{weights = TRUE}, smoother matrix of dimension
\code{nrow(xnew)} x \code{nrow(x)} containing the weights that deliver
ensemble predictions where each row gives the weight that each training
outcome received in the prediction}

 @export

 @method predict ensemble_core
}
\description{
\code{\link{ensemble}} implements an ensemble learner including the
possibility to extract the smoother matrix of linear smoothers.
}
