% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{predict.ensemble_core}
\alias{predict.ensemble_core}
\title{Prediction of fitted values for ensemble}
\usage{
\method{predict}{ensemble_core}(
  object,
  ml,
  X_tr,
  Y_tr,
  X_te,
  quiet = TRUE,
  pb = NULL,
  pb_cf = NULL,
  pb_cv = NULL,
  pb_np = NULL,
  ...
)
}
\arguments{
\item{object}{List of fitted methods (from \code{\link{ensemble_core}}}

\item{ml}{List of raw methods built via \code{\link{create_method}}}

\item{X_tr}{Covariate matrix of training sample}

\item{Y_tr}{Vector of outcomes of training sample}

\item{X_te}{Covariate matrix of test sample}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}

\item{...}{Ignore unused arguments}
}
\value{
Returns matrix of dimension \code{nrow(X_te)} X \code{length(ml)}
that contains the predictions of fitted values for each method in the ensemble.
}
\description{
Predicts fitted values for some (new) covariate matrix based on the fully
trained models of the ensemble model from \code{\link{ensemble_core}}.
}
