% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{predict.ensemble_core}
\alias{predict.ensemble_core}
\title{Prediction of fitted values for ensemble}
\usage{
\method{predict}{ensemble_core}(
  object,
  methods,
  X_tr,
  Y_tr,
  X_te,
  quiet = TRUE,
  pb = NULL,
  pb_cf = NULL,
  pb_cv = NULL,
  pb_np = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"ensemble_core"} returned by \code{\link{ensemble_core}}.}

\item{methods}{List of raw methods created with \code{\link{create_method}}.}

\item{X_tr}{Covariate matrix of the training sample.}

\item{Y_tr}{Vector of outcomes of the training sample.}

\item{X_te}{Covariate matrix of the test sample.}

\item{quiet}{Logical. If \code{FALSE}, progress output is printed to the console.}

\item{pb}{A progress bar object to track overall computation progress.}

\item{pb_cf}{Integer indicating the current cross-fitting fold 
(for progress bar updates).}

\item{pb_cv}{Integer indicating the current cross-validation fold 
(for progress bar updates).}

\item{pb_np}{String indicating the current nuisance parameter being 
estimated or predicted (for progress bar updates).}

\item{...}{Ignored additional arguments.}
}
\value{
A matrix of dimension \code{nrow(X_te)} x \code{length(methods)} containing 
predictions from each method in the ensemble. If learners return multi-column predictions, 
an array of dimension \code{nrow(X_te)} x \code{ncol(m_pred)} x \code{length(methods)} is returned.
}
\description{
Generate predictions for (new) covariates based on trained ensemble models 
from \code{\link{ensemble_core}}.
}
\keyword{internal}
