% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_short.R
\name{predict.ensemble_short}
\alias{predict.ensemble_short}
\title{Prediction of fitted values for short-stacked ensemble learner}
\usage{
\method{predict}{ensemble_short}(object, w = NULL, ...)
}
\arguments{
\item{object}{Short-stacked ensemble learner from \code{\link{ensemble_short}}}

\item{w}{Ensemble weights to aggregate predictions from different learners (optional).
Needs to be a vector of length \code{ncol(object$fit_cv)}.}

\item{...}{Additional arguments.}
}
\value{
Vector containing ensemble predictions.
}
\description{
Returns prediction of fitted values for a fitted and short-stacked ensemble
learner created by \code{\link{ensemble_short}}.
}
\keyword{internal}
