% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.forest_grf_fit}
\alias{predict.forest_grf_fit}
\title{Predict method for grf random forest models}
\usage{
\method{predict}{forest_grf_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{forest_grf_fit} from
\code{\link{forest_grf_fit}}.}

\item{Xnew}{Numeric matrix of covariates for test sample. If \code{NULL},
predictions are generated for the training data.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of predicted values.
}
\description{
Generates predictions from a fitted random forest model
estimated using the \code{grf} package.
}
\keyword{internal}
