% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.knn_fit}
\alias{predict.knn_fit}
\title{Prediction of fitted values based on the k-Nearest-Neighbor algorithm}
\usage{
\method{predict}{knn_fit}(arguments, x, y, xnew = NULL, ...)
}
\arguments{
\item{arguments}{Output of \code{\link{knn_fit}}}

\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}

\item{xnew}{Covariate matrix of test sample}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Predictions by k-Nearest-Neighbor algorithm.
Note that if no \code{k} is explicitly specified, this evaluates to
\code{k = 10} as default value.
}
\keyword{internal}
