% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.knn_fit}
\alias{predict.knn_fit}
\title{Predict method for k-Nearest Neighbors}
\usage{
\method{predict}{knn_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{knn_fit}} of class \code{knn_fit}.
List of arguments to be used in k-NN prediction. Typically
includes \code{k}. (number of neighbors). If no \code{k} is
specified, defaults to \code{k = 10}.}

\item{Xnew}{Numeric matrix of covariates for test sample. If \code{NULL},
predictions are generated for the training data.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of predicted values.
}
\description{
Generates predictions using the k-Nearest Neighbors algorithm with internal
Euclidean distance calculation, following the same approach as the
\code{Distance_for_KNN_test()} function of the \code{FastKNN} package.
}
\keyword{internal}
