% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{predict.knn_prop_fit}
\alias{predict.knn_prop_fit}
\title{Prediction based on k-Nearest Neighbor model.}
\usage{
\method{predict}{knn_prop_fit}(knn_fit, X, Y, Xnew = NULL, weights = FALSE)
}
\arguments{
\item{knn_fit}{Output of \code{\link{kknn}} or \code{\link{knn_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{weights}{Always FALSE}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for Xnew}
\item{weights}{Not supported for propensity score estimation}
}
\description{
Prediction based on k-Nearest Neighbor model.
}
\keyword{internal}
