% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.logit_fit}
\alias{predict.logit_fit}
\title{Predict method for \code{glmnet} logistic regression fits}
\usage{
\method{predict}{logit_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{logit_fit}}.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}

\item{...}{Ignored additional arguments.}
}
\value{
A data frame of predicted class probabilities.
  \describe{
    \item{Binary outcome}{A two-column data frame with probabilities for each class.}
    \item{Multiclass outcome}{A data frame with one column per class, containing the predicted probabilities.}
  }
}
\description{
Generates predictions from fitted logistic regression models.
}
\keyword{internal}
