% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{predict.logit_nnet_fit}
\alias{predict.logit_nnet_fit}
\title{Prediction based on (multinomial) logistic regression.}
\usage{
\method{predict}{logit_nnet_fit}(multinom_fit, X, Y, Xnew = NULL, weights = FALSE)
}
\arguments{
\item{multinom_fit}{Output of \code{\link{nnet}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{weights}{Always FALSE as}
}
\value{
Returns list containing:
\item{prediction}{matrix of predictions for Xnew}
\item{weights}{Not supported for propensity score estimation}
}
\description{
Prediction based on (multinomial) logistic regression.
}
\keyword{internal}
