% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.mean_fit}
\alias{predict.mean_fit}
\title{Predict method for arithmetic mean model}
\usage{
\method{predict}{mean_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{mean_fit} from \code{\link{mean_fit}}.}

\item{Xnew}{Covariate matrix for predictions. Only \code{nrow(Xnew)} is used.}

\item{...}{Ignored additional arguments.}
}
\value{
A numeric vector of predictions, all equal to the trained mean.
}
\description{
Returns a constant prediction for all observations, equal to the mean of the
training sample. Covariates \code{Xnew} are ignored.
}
\keyword{internal}
