% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{predict.nb_gaussian_fit}
\alias{predict.nb_gaussian_fit}
\title{Prediction based on Gaussian Naive Bayes model.}
\usage{
\method{predict}{nb_gaussian_fit}(nb_gaussian_fit, X, Y, Xnew = NULL, weights = FALSE)
}
\arguments{
\item{nb_gaussian_fit}{Output of \code{\link{gaussian_naive_bayes}} or \code{\link{nb_gaussian_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{weights}{Always FALSE as}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for Xnew}
\item{weights}{Not supported for propensity score estimation}
}
\description{
Prediction based on Gaussian Naive Bayes model.
}
\keyword{internal}
