% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.ols_fit}
\alias{predict.ols_fit}
\title{Predictions based on OLS}
\usage{
\method{predict}{ols_fit}(ols_fit, x, y, xnew = NULL, weights = FALSE)
}
\arguments{
\item{ols_fit}{Output of \code{\link{ols_fit}}}

\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}

\item{xnew}{Covariate matrix of test sample}

\item{weights}{If TRUE, weights underlying the prediction for xnew calculated}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for xnew}
\item{weights}{If \code{weights = TRUE} prediction weights of dimension \code{nrow(xnew)} x \code{nrow(x)}
containing the weights that deliver predictions where each row gives the weight that each training
outcome received in the prediction for xnew.}
}
\description{
Prediction based on fitted ordinary least squares models. The method also
provides prediction weights if required.
}
\keyword{internal}
