% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.ols_fit}
\alias{predict.ols_fit}
\title{Predict method for OLS models}
\usage{
\method{predict}{ols_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ols_fit} from \code{\link{ols_fit}}.}

\item{Xnew}{Numeric matrix of covariates for test sample. If \code{NULL},
predictions are generated for training data.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of predicted values.
}
\description{
Generates predictions from a fitted ordinary least squares model.
}
\keyword{internal}
