% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_OvO_OvR.R
\name{predict.ovo_fit}
\alias{predict.ovo_fit}
\title{Predicts class probabilities using One-Vs-One (OvO) routine}
\usage{
\method{predict}{ovo_fit}(ovo_fit, X, Y, Xnew = NULL, method, parallel = FALSE, verbose = TRUE)
}
\arguments{
\item{ovo_fit}{Output list of \code{\link{ovo_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{Xnew}{Covariate matrix of test sample.}

\item{method}{Method used for fitting the binary classifiers (e.g., "logit").}

\item{parallel}{Logical. Tries to run in parallel if \code{foreach} and 
\code{doParallel} packages are available. Defaults to FALSE.}

\item{verbose}{Logical. If TRUE, messages about parallel/sequential choice are printed.}
}
\value{
Matrix of class probability predictions for \code{Xnew}
}
\description{
Predicts class probabilities using One-Vs-One (OvO) fitted classifiers
with optional parallelization (auto-registers a backend if \code{foreach} 
and \code{doParallel} packages are available).
}
\keyword{internal}
