% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.plasso_fit}
\alias{predict.plasso_fit}
\title{Predict method for post-lasso regression models}
\usage{
\method{predict}{plasso_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{plasso_fit} from \code{\link{plasso_fit}}.}

\item{Xnew}{Numeric matrix of covariates for test sample. If \code{NULL},
predictions are generated for the training data.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of predicted values.
}
\description{
Generates predictions from a fitted post-lasso regression model.
Returns predictions for either new data or the training sample.
}
\keyword{internal}
