% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.prob_forest_fit}
\alias{predict.prob_forest_fit}
\title{Predict method for Probability Forest fits}
\usage{
\method{predict}{prob_forest_fit}(prob_forest_fit, X, Y, Xnew = NULL)
}
\arguments{
\item{prob_forest_fit}{Output of \code{\link{prob_forest_fit}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}
}
\value{
A numeric matrix with one column per class, containing the predicted probabilities.
}
\description{
Generates predicted class probabilities from fitted Probability Forest models.
}
\keyword{internal}
