% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.ranger_fit}
\alias{predict.ranger_fit}
\title{Predict method for Random Forest continuous outcome fits}
\usage{
\method{predict}{ranger_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{ranger_fit}}.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}

\item{...}{Ignored additional arguments.}
}
\value{
A numeric vector of predicted continuous outcomes.
}
\description{
Generates predicted values from fitted Random Forest models for continuous outcomes.
}
\keyword{internal}
