% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{predict.ranger_fit}
\alias{predict.ranger_fit}
\title{Prediction based on the \code{\link{ranger}} Probability Forest fit.}
\usage{
\method{predict}{ranger_fit}(ranger_fit, X, Y, Xnew = NULL, weights = FALSE)
}
\arguments{
\item{ranger_fit}{Output of \code{\link{ranger}} or \code{\link{prob_forest_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{weights}{Always FALSE}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for Xnew}
\item{weights}{Not supported for propensity score estimation}
}
\description{
Prediction based on the \code{\link{ranger}} Probability Forest fit.
}
\keyword{internal}
