% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.ridge_fit}
\alias{predict.ridge_fit}
\title{Predictions based on Ridge regression}
\usage{
\method{predict}{ridge_fit}(ridge_fit, xnew, ...)
}
\arguments{
\item{ridge_fit}{Output of \code{\link{ridge_fit}}}

\item{xnew}{Covariate matrix of test sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of predictions for xnew.
}
\description{
Prediction based on fitted Ridge regression model.
}
\keyword{internal}
