% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.rlasso_fit}
\alias{predict.rlasso_fit}
\title{Predictions based on \code{hdm::rlasso} regression}
\usage{
\method{predict}{rlasso_fit}(rlasso_fit, Xnew = NULL, ...)
}
\arguments{
\item{rlasso_fit}{Output of \code{\link{rlasso_fit}}}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Prediction of fitted values (for a potentially new set of covariates Xnew)
based on a trained Lasso model from hdm package.
}
\keyword{internal}
