% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{predict.svm_fit}
\alias{predict.svm_fit}
\title{Prediction based on Support Vector Machines model.}
\usage{
\method{predict}{svm_fit}(svm_fit, X, Y, Xnew = NULL, weights = FALSE)
}
\arguments{
\item{svm_fit}{Output of \code{\link{svm_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{weights}{Always FALSE as}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for Xnew}
\item{weights}{Not supported for propensity score estimation}
}
\description{
Prediction based on Support Vector Machines model.
}
\keyword{internal}
