% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.svm_fit}
\alias{predict.svm_fit}
\title{Predict method for Support Vector Machine fits}
\usage{
\method{predict}{svm_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{svm_fit}}.}

\item{Xnew}{Covariate matrix of test sample. If \code{NULL}, uses training data.}

\item{...}{Ignored additional arguments.}
}
\value{
Predicted probabilities:
  \describe{
    \item{Binary outcome}{A numeric vector of probabilities for the positive class.}
    \item{Multiclass outcome}{A numeric matrix with one column per class.}
  }
}
\description{
Generates predicted class probabilities from fitted Support Vector Machine models.
}
\keyword{internal}
