% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.xgboost_fit}
\alias{predict.xgboost_fit}
\title{Predictions based on XGBoost}
\usage{
\method{predict}{xgboost_fit}(xgboost_fit, Xnew = NULL, ...)
}
\arguments{
\item{xgboost_fit}{Output of \code{\link{xgboost_fit}}}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Vector of fitted values.
}
\description{
Prediction of fitted values (for a potentially new set of covariates Xnew)
from an XGBoost.
}
\keyword{internal}
