% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{predict.xgboost_fit}
\alias{predict.xgboost_fit}
\title{Predict method for XGBoost models}
\usage{
\method{predict}{xgboost_fit}(xgboost_fit, Xnew = NULL, ...)
}
\arguments{
\item{xgboost_fit}{An object of class \code{xgboost_fit} from \code{\link{xgboost_fit}}.}

\item{Xnew}{Numeric matrix of covariates for test sample. If \code{NULL},
predictions are generated for the training data.}

\item{...}{Ignored additional arguments.}
}
\value{
Numeric vector of predicted values.
}
\description{
Generates predictions from a fitted XGBoost model.
}
\keyword{internal}
