% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{predict.xgboost_prop_fit}
\alias{predict.xgboost_prop_fit}
\title{Prediction based on \code{xgboost} model.}
\usage{
\method{predict}{xgboost_prop_fit}(xgboost_fit, X, Y, Xnew = NULL, weights = FALSE)
}
\arguments{
\item{xgboost_fit}{Output of \code{\link{xgboost}} or \code{\link{xgboost_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{weights}{Always FALSE as}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for Xnew}
\item{weights}{Not supported for propensity score estimation}
}
\description{
Prediction based on \code{xgboost} model.
}
\keyword{internal}
