% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_cf_mat}
\alias{prep_cf_mat}
\title{Cross-fitting fold indicators}
\usage{
prep_cf_mat(N, cf, cluster = NULL, d_mat = NULL)
}
\arguments{
\item{N}{Number of observations}

\item{cf}{Number of cross-fitting folds}

\item{cluster}{Optional vector of cluster variable if cross-fitting should account
for clusters within the data.}

\item{d_mat}{Optional logical matrix of treatment indicators (\eqn{N x T+1} with \eqn{T}
being the number of treatments). For example created by \code{\link{prep_indicator_mat}}.
If specified, cross-fitting folds will preserve the treatment ratios from full sample.
However, if cluster vector is provided, \code{d_mat} is ignored due to computational
constraints and randomization as well as feasibility issues.}
}
\value{
Logical matrix of cross-fitting folds (N x # folds).
}
\description{
Creates a matrix of binary cross-fitting fold indicators (\eqn{N \times cf})
}
\examples{
\dontrun{
N <- 1000
d_mat <- prep_indicator_mat(x = sample(3, N, replace = TRUE))
cf_mat <- prep_cf_mat(N = N, cf = 5, d_mat = d_mat)
head(cf_mat)
}

}
\keyword{internal}
