% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_cf_mat}
\alias{prep_cf_mat}
\title{Cross-fitting fold indicators}
\usage{
prep_cf_mat(n, cf, w_mat = NULL, cl = NULL)
}
\arguments{
\item{n}{Number of observations}

\item{cf}{Number of cross-fitting folds}

\item{w_mat}{Optional logical matrix of treatment indicators (n x T+1 with T
being the number of treatments).
For example created by \code{\link{prep_w_mat}}.
If specified, cross-fitting folds will preserve the treatment ratios from full sample.}

\item{cl}{Optional vector of cluster variable if cross-fitting should account for clusters.
If w_mat is provided, cluster vector is ignored due to computational and feasibility constraints
for the cross-fitting procedure to work.}
}
\value{
Logical matrix of cross-fitting folds (n x # folds).
}
\description{
\code{\link{prep_cf_mat}} creates a matrix of binary cross-fitting fold
indicators (n x # cross-folds)
}
\examples{
n = 1000
w_mat = prep_w_mat(sample(3, n, replace = TRUE))
cf_mat = prep_cf_mat(n, cf = 5, w_mat = w_mat)
head(cf_mat)

}
