% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{prob_forest_fit}
\alias{prob_forest_fit}
\title{Fit Probability Forest using \code{grf}}
\usage{
prob_forest_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{arguments}{List of arguments passed to \code{grf::probability_forest()}.}
}
\value{
An object of class \code{probability_forest} containing the fitted model.
}
\description{
Fits a Probability Forest model using the \code{grf} package.
}
\keyword{internal}
