% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{ranger_fit}
\alias{ranger_fit}
\title{Calculates Probability Forest fit using the \code{\link{ranger}} package}
\usage{
ranger_fit(
  X,
  Y,
  arguments = list(num.trees = 1000, min.node.size = 0.1 * nrow(X), mtry =
    ceiling(sqrt(ncol(X))))
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{vector of outcomes}

\item{arguments}{List of arguments passed to  \code{\link{ranger}}}
}
\value{
An object with S3 class "ranger"
}
\description{
Calculates Probability Forest fit using the \code{\link{ranger}} package
}
\keyword{internal}
