% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{ridge_fit}
\alias{ridge_fit}
\title{Fits Ridge regression}
\usage{
ridge_fit(x, y, args = list())
}
\arguments{
\item{x}{Matrix of covariates}

\item{y}{vector of outcomes}

\item{args}{List of arguments passed to \code{\link[glmnet]{glmnet}}}
}
\value{
An object with S3 class \code{\link[glmnet]{glmnet}}
}
\description{
\code{\link{ridge_fit}} estimates cross-validated Ridge regression based on
the \code{\link{glmnet}} package.
}
\keyword{internal}
