% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{ridge_fit}
\alias{ridge_fit}
\title{Fit ridge regression model}
\usage{
ridge_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{arguments}{List of arguments passed to \code{cv.glmnet}.}
}
\value{
An object of class \code{ridge_fit} containing the fitted
  \code{cv.glmnet} model.
}
\description{
Estimates ridge regression using the \code{glmnet} package.
Standardizes covariates and selects optimal lambda via
cross-validation.
}
\keyword{internal}
