% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{rlasso_fit}
\alias{rlasso_fit}
\title{Fit post-lasso regression using hdm package}
\usage{
rlasso_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{arguments}{List of arguments passed to \code{\link[hdm]{rlasso}}.}
}
\value{
An object of class \code{rlasso_fit} containing the fitted
  \code{\link[hdm]{rlasso}} model.
}
\description{
Estimates post-lasso regression under homoscedastic and heteroscedastic
non-Gaussian disturbances using the \code{hdm} package.
}
\keyword{internal}
