% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{svm_fit}
\alias{svm_fit}
\title{This function calculates the Support Vector Machines model based on the \code{e1071} package}
\usage{
svm_fit(X, Y, arguments = list(gamma = 0.1))
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{arguments}{List of arguments passed to  \code{\link{svm}}}
}
\value{
An object with S3 method for class 'svm'
}
\description{
This function calculates the Support Vector Machines model based on the \code{e1071} package
}
