% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tune_learners}
\alias{tune_learners}
\title{Tune learner hyperparameters}
\usage{
tune_learners(type = c("full_sample", "fold"), methods, X, Y)
}
\arguments{
\item{type}{Tuning type, either "\code{full_sample}" or "\code{fold}".}

\item{methods}{List of methods for ensemble estimation.}

\item{X}{Covariate matrix.}

\item{Y}{Numeric vector containing the outcome variable.}
}
\value{
Modified methods list with tuned parameters.
}
\description{
Tunes hyperparameters for specified learners in a \code{methods} list
}
\keyword{internal}
