% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{weights.ensemble}
\alias{weights.ensemble}
\title{Smoother weights from cross-validated ensemble learner}
\usage{
\method{weights}{ensemble}(object, method, X, Y, Xnew, quiet = TRUE, ...)
}
\arguments{
\item{object}{Trained ensemble object from \code{\link{ensemble}}}

\item{method}{List of raw methods built via \code{\link{create_method}}}

\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of dimension \code{nrow(Xnew)} X \code{nrow(X)} containing
ensemble smoother weights.
}
\description{
\code{\link{weights.ensemble}} extract ensemble smoother weights for some
(new) covariate matrix based on a fitted and cross-validated ensemble model from
\code{\link{ensemble}}.
}
