% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{weights.ensemble_core}
\alias{weights.ensemble_core}
\title{Ensemble smoother weights}
\usage{
\method{weights}{ensemble_core}(object, ml, x_tr, y_tr, x_te, quiet = TRUE, ...)
}
\arguments{
\item{object}{List of fitted methods (from \code{\link{ensemble_core}}}

\item{ml}{List of raw methods built via \code{\link{create_method}}}

\item{x_tr}{Covariate matrix of training sample}

\item{y_tr}{Vector of outcomes of training sample}

\item{x_te}{Covariate matrix of test sample}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console}

\item{...}{Ignore unused arguments}
}
\value{
Returns a three-dimensional array of dimension
\code{nrow(xnew)} x \code{nrow(x)} x \code{length(ml)}
containing the smoother weights that deliver predictions of each method
where each row gives the weight that each training outcome received in the
prediction of the respective test set observation.
}
\description{
Extract smoother weights for some (new) covariate matrix based on the fully
trained models of the ensemble model from \code{\link{ensemble_core}}.
}
