% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_short.R
\name{weights.ensemble_short}
\alias{weights.ensemble_short}
\title{Extraction of smoother weights for short-stacked ensemble learner}
\usage{
\method{weights}{ensemble_short}(object, ml, x, y, cf_mat, quiet = TRUE, ...)
}
\arguments{
\item{object}{Short-stacked ensemble learner from \code{\link{ensemble_short}}.}

\item{ml}{List of ML models by \code{\link{create_method}} that was used as
input for \code{\link{ensemble_short}}.}

\item{x}{Covariate matrix of training sample.}

\item{y}{Vector of outcomes of training sample.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of dimension \code{nrow(x)} x \code{nrow(x)} containing
ensemble smoother weights.
}
\description{
Extract smoother weights from short-stacked ensemble model created by
\code{\link{ensemble_short}}.
}
