% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.ensemble_short}
\alias{weights.ensemble_short}
\title{Extraction of smoother weights for short-stacked ensemble learner}
\usage{
\method{weights}{ensemble_short}(
  object,
  method,
  X,
  Y,
  subset = NULL,
  w = NULL,
  cf_mat,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{object}{Short-stacked ensemble learner from \code{\link{ensemble_short}}.}

\item{method}{List of method models by \code{\link{create_method}} that was used as
input for \code{\link{ensemble_short}}.}

\item{X}{Covariate matrix of training sample.}

\item{Y}{Vector of outcomes of training sample.}

\item{subset}{Logical vector indicating which observations to use for determining
ensemble weights. If not provided, all observations are used.}

\item{w}{Ensemble weights to aggregate predictions from different learners (optional).
Needs to be a vector of length \code{ncol(object$fit_cv)}.}

\item{cf_mat}{Logical matrix with k columns of indicators representing the different folds
(for example created by \code{\link{prep_cf_mat}}).}

\item{quiet}{If FALSE, method that is currently computed is printed into the
console.}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of dimension \code{nrow(X)} X \code{nrow(X)} containing
ensemble smoother weights.
}
\description{
Extract smoother weights from short-stacked ensemble model created by
\code{\link{ensemble_short}}.
}
