% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{weights.forest_grf_fit}
\alias{weights.forest_grf_fit}
\title{Smoother weights from random forest model}
\usage{
\method{weights}{forest_grf_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{forest_grf_fit}}.}

\item{Xnew}{Covariate matrix of test sample. If not provided, 
prediction is done for the training sample.}

\item{...}{Ignore unused arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Extract smoother weights for test sample from a random forest model.
}
\keyword{internal}
