% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.mean_fit}
\alias{weights.mean_fit}
\title{Compute smoother weights from arithmetic mean}
\usage{
\method{weights}{mean_fit}(mean_fit, Xnew, ...)
}
\arguments{
\item{mean_fit}{An object returned by \code{\link{mean_fit}}.}

\item{Xnew}{Covariate matrix of the test sample.}

\item{...}{Ignored arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Generate smoother weights for a test sample using arithmetic mean fitting of the training sample.
}
\keyword{internal}
