% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R, R/ml_wrapper.R
\name{weights.ols_fit}
\alias{weights.ols_fit}
\title{Smoother weights from OLS prediction}
\usage{
\method{weights}{ols_fit}(ols_fit, X, Xnew, ...)

\method{weights}{ols_fit}(ols_fit, X, Xnew, ...)
}
\arguments{
\item{ols_fit}{Output of \code{\link{ols_fit}}}

\item{X}{Covariate matrix of training sample}

\item{Xnew}{Covariate matrix of test sample}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of smoother weights.

Matrix of smoother weights.
}
\description{
Extract smoother weights for test sample from an OLS model.

Extract smoother weights for test sample from an OLS model.
}
\keyword{internal}
