% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R, R/ml_wrapper.R
\name{weights.plasso_fit}
\alias{weights.plasso_fit}
\title{Smoother weights from Post-Lasso prediction}
\usage{
\method{weights}{plasso_fit}(plasso_fit, xnew = NULL, ...)

\method{weights}{plasso_fit}(plasso_fit, xnew = NULL, ...)
}
\arguments{
\item{plasso_fit}{Output of \code{\link{plasso_fit}}}

\item{xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of smoother weights.

Matrix of smoother weights.
}
\description{
Extract smoother weights for test sample from a Post-Lasso regression model.

Extract smoother weights for test sample from a Post-Lasso regression model.
}
\keyword{internal}
