% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{weights.ridge_fit}
\alias{weights.ridge_fit}
\title{Smoother weights from Ridge regression prediction}
\usage{
\method{weights}{ridge_fit}(ridge_fit, x, y, xnew = NULL, ...)
}
\arguments{
\item{ridge_fit}{Output of \code{\link{ridge_fit}}}

\item{x}{Covariate matrix of training sample}

\item{y}{Vector of outcomes of training sample}

\item{xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments}
}
\value{
Matrix of smoother weights.
}
\description{
Extract smoother weights for test sample from a Ridge regression model.
}
\keyword{internal}
