% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{weights.ridge_fit}
\alias{weights.ridge_fit}
\title{Compute smoother weights from ridge regression}
\usage{
\method{weights}{ridge_fit}(ridge_fit, X, Y, Xnew = NULL, ...)
}
\arguments{
\item{ridge_fit}{An object returned by \code{\link{ridge_fit}}.}

\item{X}{Covariate matrix of the training sample.}

\item{Y}{Vector of outcomes of the training sample.}

\item{Xnew}{Optional covariate matrix of the test sample. If not provided, predictions are computed for the training sample.}

\item{...}{Ignored arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Generate smoother weights for a test sample from a ridge regression model.
}
\keyword{internal}
