% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper.R
\name{xgboost_fit}
\alias{xgboost_fit}
\title{Fit XGBoost model}
\usage{
xgboost_fit(X, Y, arguments = list())
}
\arguments{
\item{X}{Numeric matrix of covariates for training sample.}

\item{Y}{Numeric vector of outcomes for training sample.}

\item{arguments}{List of arguments passed to \code{xgb.train}.}
}
\value{
An object of class \code{xgboost_fit} containing the fitted
  \code{xgb.Booster} model.
}
\description{
Fits an XGBoost model using the \code{xgboost} package.
}
\details{
For successful smoother extraction in \pkg{OutcomeWeights}, certain parameter
restrictions are recommended. If not specified by the user, the following
values are suggested:
\itemize{
  \item \code{max_delta_step = 0}
  \item \code{base_score = mean(Y)} 
  \item \code{subsample = 1} 
  \item \code{alpha = 0}
}
User-specified values in \code{arguments} take precedence over suggestions.
}
\keyword{internal}
