% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wrapper_prop.R
\name{xgboost_prop_fit}
\alias{xgboost_prop_fit}
\title{This function calculates the xgboost model based on the \code{\link{xgboost}} package}
\usage{
xgboost_prop_fit(X, Y, arguments = list(nrounds = 40, eta = 0.01))
}
\arguments{
\item{X}{Covariate matrix of training sample}

\item{Y}{Vector of outcomes of training sample}

\item{arguments}{List of arguments passed to  \code{\link{xgboost}}}
}
\value{
An object with S3 method for class 'xgb.Booster'
}
\description{
This function calculates the xgboost model based on the \code{\link{xgboost}} package
}
